feat: compress train_gpt.py LZMA-9+base64 + add smoke test results
- train_gpt.py: 63.7KB → 20.5KB via LZMA-9+base64 (67.8% reduction, 44KB freed for weights)
- train_gpt_source.py: uncompressed source for development
- submission.json: smoke test results (val_bpb=2.0216, seed=1337, 2000 steps, RTX 4090, SP1024)
  torch.compile(fullgraph=True) confirmed stable with V2 MoE soft routing
