+++ records/track_10min_16mb/anticipatory-transformer/train_gpt.py
@@ -1,1475 +1,2 @@
-"""
-Anticipatory Transformer V2 — Parameter Golf Submission
-========================================================
-Building on SOTA techniques (Int5/Int6 mixed quantization, BigramHash,
-SmearGate, OrthoInit, MuonWD, SWA, sliding window eval) with three
-innovations from anticipation geometry:
-
-1. MoE Trajectory Routing — 10 specialist expert networks derive per-position
-   routing from 7 anticipation scalars (commitment, uncertainty, transition
-   pressure, recovery margin, phase stiffness, novelty, stability). Each
-   expert learns a different per-head modulation pattern. Zero-parameter
-   routing via learned centroids. Validated: -7.4% perplexity on wikitext-2.
-2. Entropy-Weighted Training Loss — upweight high-entropy tokens during
-   training. Focuses gradient on informative positions.
-3. 7-Scalar Trajectory Extraction — compute commitment, uncertainty, and
-   5 other geometric scalars from hidden state dynamics. Necessary and
-   sufficient for convergence (validated on 291K ASR data: without scalars,
-   baseline collapses to 100% CER).
-
-Key result: MoE inscription routing with 10 behavioral-category experts
-gives -7.4% perplexity vs control (wikitext-2, 4-way ablation). Hard
-routing beats soft embedding (-6.8%). Scalars alone give -4.4%.
-"""
-
-from __future__ import annotations
-
-import copy
-import glob
-import io
-import math
-import os
-import random
-import subprocess
-import sys
-import time
-import uuid
-import zlib
-from pathlib import Path
-
-try:
-    import zstandard
-    _COMPRESSOR = "zstd"
-except ImportError:
-    _COMPRESSOR = "zlib"
-
-import numpy as np
-import sentencepiece as spm
-import torch
-import torch.distributed as dist
-import torch.nn.functional as F
-from torch import Tensor, nn
-from torch.nn.parallel import DistributedDataParallel as DDP
-
-# -----------------------------
-# HYPERPARAMETERS
-# -----------------------------
-
-class Hyperparameters:
-    data_path = os.environ.get("DATA_PATH", "./data/datasets/fineweb10B_sp1024")
-    train_files = os.path.join(data_path, "fineweb_train_*.bin")
-    val_files = os.path.join(data_path, "fineweb_val_*.bin")
-    tokenizer_path = os.environ.get("TOKENIZER_PATH", "./data/tokenizers/fineweb_1024_bpe.model")
-    run_id = os.environ.get("RUN_ID", str(uuid.uuid4()))
-    seed = int(os.environ.get("SEED", 42))
-
-    val_batch_size = int(os.environ.get("VAL_BATCH_SIZE", 524_288))
-    val_loss_every = int(os.environ.get("VAL_LOSS_EVERY", 500))
-    train_log_every = int(os.environ.get("TRAIN_LOG_EVERY", 100))
-
-    iterations = int(os.environ.get("ITERATIONS", 20000))
-    warmdown_iters = int(os.environ.get("WARMDOWN_ITERS", 3000))
-    warmup_steps = int(os.environ.get("WARMUP_STEPS", 20))
-    train_batch_tokens = int(os.environ.get("TRAIN_BATCH_TOKENS", 786_432))
-    train_seq_len = int(os.environ.get("TRAIN_SEQ_LEN", 2048))
-    max_wallclock_seconds = float(os.environ.get("MAX_WALLCLOCK_SECONDS", 600.0))
-    qk_gain_init = float(os.environ.get("QK_GAIN_INIT", 1.5))
-
-    vocab_size = int(os.environ.get("VOCAB_SIZE", 1024))
-    num_layers = int(os.environ.get("NUM_LAYERS", 10))
-    num_kv_heads = int(os.environ.get("NUM_KV_HEADS", 4))
-    model_dim = int(os.environ.get("MODEL_DIM", 512))
-    num_heads = int(os.environ.get("NUM_HEADS", 8))
-    mlp_mult = float(os.environ.get("MLP_MULT", 3.0))
-    tie_embeddings = bool(int(os.environ.get("TIE_EMBEDDINGS", "1")))
-    rope_base = float(os.environ.get("ROPE_BASE", 10000.0))
-    logit_softcap = float(os.environ.get("LOGIT_SOFTCAP", 30.0))
-
-    embed_lr = float(os.environ.get("EMBED_LR", 0.6))
-    head_lr = float(os.environ.get("HEAD_LR", 0.008))
-    tied_embed_lr = float(os.environ.get("TIED_EMBED_LR", 0.03))
-    tied_embed_init_std = float(os.environ.get("TIED_EMBED_INIT_STD", 0.005))
-    matrix_lr = float(os.environ.get("MATRIX_LR", 0.02))
-    scalar_lr = float(os.environ.get("SCALAR_LR", 0.02))
-    muon_momentum = float(os.environ.get("MUON_MOMENTUM", 0.99))
-    muon_backend_steps = int(os.environ.get("MUON_BACKEND_STEPS", 5))
-    muon_momentum_warmup_start = float(os.environ.get("MUON_MOMENTUM_WARMUP_START", 0.92))
-    muon_momentum_warmup_steps = int(os.environ.get("MUON_MOMENTUM_WARMUP_STEPS", 1500))
-    beta1 = float(os.environ.get("BETA1", 0.9))
-    beta2 = float(os.environ.get("BETA2", 0.95))
-    adam_eps = float(os.environ.get("ADAM_EPS", 1e-8))
-    grad_clip_norm = float(os.environ.get("GRAD_CLIP_NORM", 0.3))
-    weight_decay = float(os.environ.get("WEIGHT_DECAY", 0.04))
-
-    eval_stride = int(os.environ.get("EVAL_STRIDE", 64))
-    eval_batch_seqs = int(os.environ.get("EVAL_BATCH_SEQS", 32))
-
-    bigram_vocab_size = int(os.environ.get("BIGRAM_VOCAB_SIZE", 10240))
-    bigram_dim = int(os.environ.get("BIGRAM_DIM", 128))
-
-    # SWA + EMA hybrid: SWA every N steps, EMA smoothing on the SWA accumulator
-    swa_enabled = bool(int(os.environ.get("SWA_ENABLED", "1")))
-    swa_start_frac = float(os.environ.get("SWA_START_FRAC", 0.4))
-    swa_every = int(os.environ.get("SWA_EVERY", 50))
-
-    # Entropy-weighted loss: upweight hard tokens during training
-    entropy_loss_enabled = bool(int(os.environ.get("ENTROPY_LOSS_ENABLED", "1")))
-    entropy_loss_alpha = float(os.environ.get("ENTROPY_LOSS_ALPHA", 0.15))
-
-    # Trajectory-scaled attention: commitment gating (V1, kept for ablation)
-    trajectory_attn_enabled = bool(int(os.environ.get("TRAJECTORY_ATTN_ENABLED", "1")))
-    trajectory_attn_strength = float(os.environ.get("TRAJECTORY_ATTN_STRENGTH", 0.1))
-
-    # MoE trajectory routing (V2): 10 experts, 7 scalars, shared across layers
-    moe_trajectory_enabled = bool(int(os.environ.get("MOE_TRAJECTORY_ENABLED", "1")))
-    moe_n_experts = int(os.environ.get("MOE_N_EXPERTS", 10))
-    moe_n_scalars = int(os.environ.get("MOE_N_SCALARS", 7))
-    moe_ortho_weight = float(os.environ.get("MOE_ORTHO_WEIGHT", 0.01))
-
-# -----------------------------
-# MUON OPTIMIZER
-# -----------------------------
-
-def zeropower_via_newtonschulz5(G: Tensor, steps: int = 10, eps: float = 1e-7) -> Tensor:
-    a, b, c = (3.4445, -4.7750, 2.0315)
-    X = G.bfloat16()
-    X /= X.norm() + eps
-    transposed = G.size(0) > G.size(1)
-    if transposed:
-        X = X.T
-    for _ in range(steps):
-        A = X @ X.T
-        B = b * A + c * A @ A
-        X = a * X + B @ X
-    return X.T if transposed else X
-
-
-class Muon(torch.optim.Optimizer):
-    def __init__(self, params, lr: float, momentum: float, backend_steps: int, nesterov: bool = True, weight_decay: float = 0.0):
-        super().__init__(
-            params,
-            dict(lr=lr, momentum=momentum, backend_steps=backend_steps, nesterov=nesterov, weight_decay=weight_decay),
-        )
-
-    @torch.no_grad()
-    def step(self, closure=None):
-        loss = None
-        if closure is not None:
-            with torch.enable_grad():
-                loss = closure()
-        distributed = dist.is_available() and dist.is_initialized()
-        world_size = dist.get_world_size() if distributed else 1
-        rank = dist.get_rank() if distributed else 0
-
-        for group in self.param_groups:
-            params = group["params"]
-            if not params:
-                continue
-            lr = group["lr"]
-            momentum = group["momentum"]
-            backend_steps = group["backend_steps"]
-            nesterov = group["nesterov"]
-
-            total_params = sum(int(p.numel()) for p in params)
-            updates_flat = torch.zeros(total_params, device=params[0].device, dtype=torch.bfloat16)
-
-            curr = 0
-            for i, p in enumerate(params):
-                if i % world_size == rank and p.grad is not None:
-                    g = p.grad
-                    state = self.state[p]
-                    if "momentum_buffer" not in state:
-                        state["momentum_buffer"] = torch.zeros_like(g)
-                    buf = state["momentum_buffer"]
-                    buf.mul_(momentum).add_(g)
-                    if nesterov:
-                        g = g.add(buf, alpha=momentum)
-                    g = zeropower_via_newtonschulz5(g, steps=backend_steps)
-                    g *= max(1, g.size(0) / g.size(1)) ** 0.5
-                    updates_flat[curr : curr + p.numel()] = g.reshape(-1)
-                curr += p.numel()
-
-            if distributed:
-                dist.all_reduce(updates_flat, op=dist.ReduceOp.SUM)
-
-            wd = group.get("weight_decay", 0.0)
-            curr = 0
-            for p in params:
-                g = updates_flat[curr : curr + p.numel()].view_as(p).to(dtype=p.dtype)
-                if wd > 0:
-                    p.data.mul_(1.0 - lr * wd)
-                p.add_(g, alpha=-lr)
-                curr += p.numel()
-        return loss
-
-
-# -----------------------------
-# TOKENIZER-AGNOSTIC EVALUATION
-# -----------------------------
-
-def build_sentencepiece_luts(
-    sp: spm.SentencePieceProcessor, vocab_size: int, device: torch.device
-) -> tuple[Tensor, Tensor, Tensor]:
-    sp_vocab_size = int(sp.vocab_size())
-    table_size = max(sp_vocab_size, vocab_size)
-    base_bytes_np = np.zeros((table_size,), dtype=np.int16)
-    has_leading_space_np = np.zeros((table_size,), dtype=np.bool_)
-    is_boundary_token_np = np.ones((table_size,), dtype=np.bool_)
-    for token_id in range(sp_vocab_size):
-        if sp.is_control(token_id) or sp.is_unknown(token_id) or sp.is_unused(token_id):
-            continue
-        is_boundary_token_np[token_id] = False
-        if sp.is_byte(token_id):
-            base_bytes_np[token_id] = 1
-            continue
-        piece = sp.id_to_piece(token_id)
-        if piece.startswith("\u2581"):
-            has_leading_space_np[token_id] = True
-            piece = piece[1:]
-        base_bytes_np[token_id] = len(piece.encode("utf-8"))
-    return (
-        torch.tensor(base_bytes_np, dtype=torch.int16, device=device),
-        torch.tensor(has_leading_space_np, dtype=torch.bool, device=device),
-        torch.tensor(is_boundary_token_np, dtype=torch.bool, device=device),
-    )
-
-
-def load_validation_tokens(pattern: str, seq_len: int) -> Tensor:
-    files = [Path(p) for p in sorted(glob.glob(pattern))]
-    if not files:
-        raise FileNotFoundError(f"No files found for pattern: {pattern}")
-    tokens = torch.cat([load_data_shard(file) for file in files]).contiguous()
-    usable = ((tokens.numel() - 1) // seq_len) * seq_len
-    if usable <= 0:
-        raise ValueError(f"Validation split is too short for TRAIN_SEQ_LEN={seq_len}")
-    return tokens[: usable + 1]
-
-
-def eval_val(
-    args: Hyperparameters,
-    model: nn.Module,
-    rank: int,
-    world_size: int,
-    device: torch.device,
-    grad_accum_steps: int,
-    val_tokens: Tensor,
-    base_bytes_lut: Tensor,
-    has_leading_space_lut: Tensor,
-    is_boundary_token_lut: Tensor,
-) -> tuple[float, float]:
-    local_batch_tokens = args.val_batch_size // (world_size * grad_accum_steps)
-    if local_batch_tokens < args.train_seq_len:
-        raise ValueError(
-            "VAL_BATCH_SIZE must provide at least one sequence per rank; "
-            f"got VAL_BATCH_SIZE={args.val_batch_size}, WORLD_SIZE={world_size}, "
-            f"GRAD_ACCUM_STEPS={grad_accum_steps}, TRAIN_SEQ_LEN={args.train_seq_len}"
-        )
-    local_batch_seqs = local_batch_tokens // args.train_seq_len
-    total_seqs = (val_tokens.numel() - 1) // args.train_seq_len
-    seq_start = (total_seqs * rank) // world_size
-    seq_end = (total_seqs * (rank + 1)) // world_size
-    val_loss_sum = torch.zeros((), device=device, dtype=torch.float64)
-    val_token_count = torch.zeros((), device=device, dtype=torch.float64)
-    val_byte_count = torch.zeros((), device=device, dtype=torch.float64)
-    model.eval()
-    with torch.inference_mode():
-        for batch_seq_start in range(seq_start, seq_end, local_batch_seqs):
-            batch_seq_end = min(batch_seq_start + local_batch_seqs, seq_end)
-            raw_start = batch_seq_start * args.train_seq_len
-            raw_end = batch_seq_end * args.train_seq_len + 1
-            local = val_tokens[raw_start:raw_end].to(device=device, dtype=torch.int64, non_blocking=True)
-            x = local[:-1].reshape(-1, args.train_seq_len)
-            y = local[1:].reshape(-1, args.train_seq_len)
-            with torch.autocast(device_type="cuda", dtype=torch.bfloat16, enabled=True):
-                batch_loss = model(x, y).detach()
-            batch_token_count = float(y.numel())
-            val_loss_sum += batch_loss.to(torch.float64) * batch_token_count
-            val_token_count += batch_token_count
-            prev_ids = x.reshape(-1)
-            tgt_ids = y.reshape(-1)
-            token_bytes = base_bytes_lut[tgt_ids].to(dtype=torch.int16)
-            token_bytes += (has_leading_space_lut[tgt_ids] & ~is_boundary_token_lut[prev_ids]).to(dtype=torch.int16)
-            val_byte_count += token_bytes.to(torch.float64).sum()
-    if dist.is_available() and dist.is_initialized():
-        dist.all_reduce(val_loss_sum, op=dist.ReduceOp.SUM)
-        dist.all_reduce(val_token_count, op=dist.ReduceOp.SUM)
-        dist.all_reduce(val_byte_count, op=dist.ReduceOp.SUM)
-    val_loss = val_loss_sum / val_token_count
-    bits_per_token = val_loss.item() / math.log(2.0)
-    tokens_per_byte = val_token_count.item() / val_byte_count.item()
-    model.train()
-    return float(val_loss.item()), float(bits_per_token * tokens_per_byte)
-
-
-# -----------------------------
-# POST-TRAINING QUANTIZATION (INT8 legacy + INT6 mixed)
-# -----------------------------
-
-CONTROL_TENSOR_NAME_PATTERNS = tuple(
-    pattern
-    for pattern in os.environ.get(
-        "CONTROL_TENSOR_NAME_PATTERNS",
-        "attn_scale,attn_scales,mlp_scale,mlp_scales,resid_mix,resid_mixes,q_gain,skip_weight,skip_weights,smear,bigram.scale,traj_gate",
-    ).split(",")
-    if pattern
-)
-FP16_KEEP_NAME_PATTERNS = tuple(
-    pattern
-    for pattern in os.environ.get("FP16_KEEP_NAME_PATTERNS", "tok_emb,blocks.8.attn.c_k,blocks.9.attn.c_k").split(",")
-    if pattern
-)
-INT8_KEEP_FLOAT_FP32_NAME_PATTERNS = tuple(
-    pattern
-    for pattern in os.environ.get(
-        "INT8_KEEP_FLOAT_FP32_NAME_PATTERNS",
-        ",".join(CONTROL_TENSOR_NAME_PATTERNS),
-    ).split(",")
-    if pattern
-)
-INT8_KEEP_FLOAT_MAX_NUMEL = 65_536
-INT8_KEEP_FLOAT_STORE_DTYPE = torch.float16
-INT8_PER_ROW_SCALE_DTYPE = torch.float16
-INT8_CLIP_PERCENTILE = 99.99984
-INT8_CLIP_Q = INT8_CLIP_PERCENTILE / 100.0
-
-def tensor_nbytes(t: Tensor) -> int:
-    return int(t.numel()) * int(t.element_size())
-
-def quantize_float_tensor(t: Tensor) -> tuple[Tensor, Tensor]:
-    t32 = t.float()
-    if t32.ndim == 2:
-        clip_abs = (
-            torch.quantile(t32.abs(), INT8_CLIP_Q, dim=1)
-            if t32.numel()
-            else torch.empty((t32.shape[0],), dtype=torch.float32)
-        )
-        clipped = torch.maximum(torch.minimum(t32, clip_abs[:, None]), -clip_abs[:, None])
-        scale = (clip_abs / 127.0).clamp_min(1.0 / 127.0)
-        q = torch.clamp(torch.round(clipped / scale[:, None]), -127, 127).to(torch.int8).contiguous()
-        return q, scale.to(dtype=INT8_PER_ROW_SCALE_DTYPE).contiguous()
-    clip_abs = float(torch.quantile(t32.abs().flatten(), INT8_CLIP_Q).item()) if t32.numel() else 0.0
-    scale = torch.tensor(clip_abs / 127.0 if clip_abs > 0 else 1.0, dtype=torch.float32)
-    q = torch.clamp(torch.round(torch.clamp(t32, -clip_abs, clip_abs) / scale), -127, 127).to(torch.int8).contiguous()
-    return q, scale
-
-
-def _classify_param(name: str) -> str:
-    if "tok_emb" in name or "lm_head" in name:
-        return "embed"
-    if ".mlp." in name:
-        return "mlp"
-    if "bigram" in name:
-        return "bigram"
-    if ".attn." in name or (".proj." in name and ".mlp." not in name):
-        return "attn"
-    return "other"
-
-def quantize_intN_per_row(t: Tensor, clip_range: int = 31) -> tuple[Tensor, Tensor]:
-    t32 = t.float()
-    if t32.ndim == 2:
-        row_max = t32.abs().amax(dim=1)
-        scale = (row_max / clip_range).clamp_min(1e-12).to(torch.float16)
-        scale = scale.clamp_min(torch.finfo(torch.float16).tiny)
-        q = torch.clamp(torch.round(t32 / scale.float()[:, None]), -(clip_range+1), clip_range).to(torch.int8)
-        return q, scale
-    amax = t32.abs().max().item()
-    scale = torch.tensor(max(amax / clip_range, 1e-12), dtype=torch.float16)
-    q = torch.clamp(torch.round(t32 / scale.float()), -(clip_range+1), clip_range).to(torch.int8)
-    return q, scale
-
-def mixed_quantize_int6(state_dict: dict[str, Tensor], int6_cats: set[str]):
-    result: dict[str, Tensor] = {}
-    meta: dict[str, object] = {}
-    for name, tensor in state_dict.items():
-        t = tensor.detach().cpu().contiguous()
-        cat = _classify_param(name)
-        if not t.is_floating_point() or t.numel() <= 8192:
-            result[name] = t.to(torch.float16) if t.is_floating_point() else t
-            meta[name] = "passthrough"
-            continue
-        if any(p in name for p in CONTROL_TENSOR_NAME_PATTERNS):
-            result[name] = t.float()
-            meta[name] = "passthrough_ctrl"
-            continue
-        if any(pattern in name for pattern in FP16_KEEP_NAME_PATTERNS):
-            result[name] = t.to(dtype=torch.float16).contiguous()
-            meta[name] = "passthrough_fp16"
-            continue
-        if cat in int6_cats and t.ndim >= 1:
-            clip = 15 if cat == "mlp" else 31  # int5 for MLP, int6 for attention
-            q, s = quantize_intN_per_row(t, clip_range=clip)
-            result[name + ".q"] = q
-            result[name + ".scale"] = s
-            meta[name] = {"type": f"int{5 if cat == 'mlp' else 6}"}
-        else:
-            q, s = quantize_float_tensor(t)
-            result[name + ".q"] = q
-            result[name + ".scale"] = s
-            meta[name] = {"type": "int8"}
-    return result, meta
-
-def dequantize_mixed_int6(result: dict[str, Tensor], meta: dict[str, object],
-                          template_sd: dict[str, Tensor]) -> dict[str, Tensor]:
-    out: dict[str, Tensor] = {}
-    for name, orig in template_sd.items():
-        info = meta[name]
-        orig_dtype = orig.dtype
-        if info in ("passthrough", "passthrough_ctrl", "passthrough_fp16"):
-            t = result[name]
-            if t.dtype == torch.float16 and orig_dtype in (torch.float32, torch.bfloat16):
-                t = t.to(orig_dtype)
-            out[name] = t
-            continue
-        q, s = result[name + ".q"], result[name + ".scale"]
-        if s.ndim > 0:
-            out[name] = (q.float() * s.float().view(q.shape[0], *([1] * (q.ndim - 1)))).to(orig_dtype)
-        else:
-            out[name] = (q.float() * float(s.item())).to(orig_dtype)
-    return out
-
-
-# -----------------------------
-# DATA LOADING
-# -----------------------------
-
-def load_data_shard(file: Path) -> Tensor:
-    header_bytes = 256 * np.dtype("<i4").itemsize
-    token_bytes = np.dtype("<u2").itemsize
-    header = np.fromfile(file, dtype="<i4", count=256)
-    if header.size != 256 or int(header[0]) != 20240520 or int(header[1]) != 1:
-        raise ValueError(f"Unexpected shard header for {file}")
-    num_tokens = int(header[2])
-    expected_size = header_bytes + num_tokens * token_bytes
-    if file.stat().st_size != expected_size:
-        raise ValueError(f"Shard size mismatch for {file}: expected {expected_size} bytes")
-    tokens_np = np.fromfile(file, dtype="<u2", count=num_tokens, offset=header_bytes)
-    if tokens_np.size != num_tokens:
-        raise ValueError(f"Short read for {file}")
-    return torch.from_numpy(tokens_np.astype(np.uint16, copy=False))
-
-
-class TokenStream:
-    def __init__(self, pattern: str):
-        self.files = [Path(p) for p in sorted(glob.glob(pattern))]
-        if not self.files:
-            raise FileNotFoundError(f"No files found for pattern: {pattern}")
-        self.file_idx = 0
-        self.tokens = load_data_shard(self.files[0])
-        self.pos = 0
-
-    def _advance_file(self) -> None:
-        self.file_idx = (self.file_idx + 1) % len(self.files)
-        self.tokens = load_data_shard(self.files[self.file_idx])
-        self.pos = 0
-
-    def take(self, n: int) -> Tensor:
-        chunks: list[Tensor] = []
-        remaining = n
-        while remaining > 0:
-            avail = self.tokens.numel() - self.pos
-            if avail <= 0:
-                self._advance_file()
-                continue
-            k = min(remaining, avail)
-            chunks.append(self.tokens[self.pos : self.pos + k])
-            self.pos += k
-            remaining -= k
-        return chunks[0] if len(chunks) == 1 else torch.cat(chunks)
-
-
-class DistributedTokenLoader:
-    def __init__(self, pattern: str, rank: int, world_size: int, device: torch.device):
-        self.rank = rank
-        self.world_size = world_size
-        self.device = device
-        self.stream = TokenStream(pattern)
-
-    def next_batch(self, global_tokens: int, seq_len: int, grad_accum_steps: int) -> tuple[Tensor, Tensor]:
-        local_tokens = global_tokens // (self.world_size * grad_accum_steps)
-        per_rank_span = local_tokens + 1
-        chunk = self.stream.take(per_rank_span * self.world_size)
-        start = self.rank * per_rank_span
-        local = chunk[start : start + per_rank_span].to(dtype=torch.int64)
-        x = local[:-1].reshape(-1, seq_len)
-        y = local[1:].reshape(-1, seq_len)
-        return x.to(self.device, non_blocking=True), y.to(self.device, non_blocking=True)
-
-
-# -----------------------------
-# TRANSFORMER MODULES
-# -----------------------------
-
-class RMSNorm(nn.Module):
-    def __init__(self, eps: float | None = None):
-        super().__init__()
-        self.eps = eps
-
-    def forward(self, x: Tensor) -> Tensor:
-        return F.rms_norm(x, (x.size(-1),), eps=self.eps)
-
-
-# -----------------------------
-# MoE TRAJECTORY ROUTING (V2)
-# -----------------------------
-
-class TrajectoryScalarExtractor(nn.Module):
-    """Extract 7 anticipation scalars from hidden state dynamics.
-
-    Uses 2D linear projections (no conv1d/3D tensors) for clean torch.compile
-    and quantization. Temporal context via diff features (parameter-free).
-    Params: dim*mid + mid + mid*7 + 7 = ~4.2K for dim=512.
-    """
-
-    def __init__(self, dim: int, n_scalars: int = 7):
-        super().__init__()
-        mid = max(dim // 64, 8)  # 8 for dim=512
-        # All 2D weights for clean int6 quantization
-        self.proj1_weight = nn.Parameter(torch.zeros(mid, dim))
-        self.proj1_bias = nn.Parameter(torch.zeros(mid))
-        self.proj2_weight = nn.Parameter(torch.zeros(n_scalars, mid))
-        self.proj2_bias = nn.Parameter(torch.zeros(n_scalars))
-        nn.init.xavier_uniform_(self.proj1_weight, gain=0.1)
-        nn.init.xavier_uniform_(self.proj2_weight, gain=0.1)
-
-    def forward(self, x: Tensor) -> Tensor:
-        # x: (batch, seq, dim) -> diff features for temporal context
-        x_diff = F.pad(x[:, 1:] - x[:, :-1], (0, 0, 1, 0))  # (batch, seq, dim)
-        h = F.gelu(F.linear(x_diff, self.proj1_weight, self.proj1_bias))  # (batch, seq, mid)
-        return torch.sigmoid(F.linear(h, self.proj2_weight, self.proj2_bias))  # (batch, seq, 7)
-
-
-class MoETrajectoryRouter(nn.Module):
-    """Soft MoE routing: 10 experts, fully differentiable, compile-safe.
-
-    Fixes from adversarial review:
-    - Soft routing via softmax (not argmin) -> gradients flow to centroids
-    - Manual squared distance (not torch.cdist) -> fullgraph=True safe
-    - All 2D weight matrices (not 3D) -> clean int6 quantization
-    - Centroids init in [0,1] (not randn) -> matches sigmoid-bounded scalars
-    Params: 10*7 + 7*160 + 10*16 + 16*80 + 10*8 = ~2.6K
-    """
-
-    def __init__(self, n_scalars: int, n_heads: int, n_experts: int = 10):
-        super().__init__()
-        self.n_experts = n_experts
-        self.n_heads = n_heads
-        hidden = max(n_scalars * 2, 16)  # 16
-        self._hidden = hidden
-
-        # Centroids in [0,1] to match sigmoid scalars (no weight decay)
-        self.centroids = nn.Parameter(torch.rand(n_experts, n_scalars))
-
-        # Expert layer 1: shared 2D weight (n_experts*hidden, n_scalars)
-        self.w1 = nn.Linear(n_scalars, n_experts * hidden, bias=False)
-        self.b1 = nn.Parameter(torch.zeros(n_experts * hidden))
-        # Expert layer 2: 2D weight (n_experts*n_heads, hidden)
-        self.w2_weight = nn.Parameter(torch.zeros(n_experts * n_heads, hidden))
-        self.b2 = nn.Parameter(torch.zeros(n_experts * n_heads))
-        nn.init.xavier_uniform_(self.w1.weight, gain=0.1)
-        nn.init.xavier_uniform_(self.w2_weight, gain=0.1)
-
-    def forward(self, scalars: Tensor) -> Tensor:
-        batch, seq, n_s = scalars.shape
-        E, H = self.n_experts, self._hidden
-
-        # Soft routing: squared distance, no torch.cdist or argmin
-        diff = scalars.unsqueeze(2) - self.centroids  # (B, S, E, 7)
-        sq_dists = (diff * diff).sum(dim=-1)  # (B, S, E)
-        weights = F.softmax(-sq_dists, dim=-1)  # (B, S, E)
-
-        # All expert outputs in parallel via 2D matmuls
-        flat = scalars.reshape(-1, n_s)  # (B*S, 7)
-        h1 = self.w1(flat) + self.b1  # (B*S, E*H)
-        h1 = F.gelu(h1).reshape(-1, E, H)  # (B*S, E, H)
-
-        # Layer 2 via einsum with 2D weight reshaped to (E, n_heads, H)
-        w2 = self.w2_weight.reshape(E, self.n_heads, H)
-        h2 = torch.einsum('beh,eoh->beo', h1, w2)  # (B*S, E, n_heads)
-        h2 = h2 + self.b2.reshape(E, self.n_heads)  # broadcast
-
-        # Weighted combination across experts
-        w_flat = weights.reshape(-1, E)  # (B*S, E)
-        out = torch.einsum('be,beo->bo', w_flat, h2)  # (B*S, n_heads)
-
-        modulation = 0.8 + 0.4 * torch.sigmoid(out)
-        return modulation.reshape(batch, seq, self.n_heads).permute(0, 2, 1).unsqueeze(-1)
-
-    def orthogonality_penalty(self) -> Tensor:
-        c = F.normalize(self.centroids, dim=-1)
-        gram = c @ c.T
-        eye = torch.eye(self.n_experts, device=gram.device)
-        return ((gram - eye) ** 2).sum()
-
-
-class CastedLinear(nn.Linear):
-    def forward(self, x: Tensor) -> Tensor:
-        w = self.weight.to(x.dtype)
-        bias = self.bias.to(x.dtype) if self.bias is not None else None
-        return F.linear(x, w, bias)
-
-
-def restore_low_dim_params_to_fp32(module: nn.Module) -> None:
-    with torch.no_grad():
-        for name, param in module.named_parameters():
-            if (param.ndim < 2 or any(pattern in name for pattern in CONTROL_TENSOR_NAME_PATTERNS)) and param.dtype != torch.float32:
-                param.data = param.data.float()
-
-
-class Rotary(nn.Module):
-    def __init__(self, dim: int, base: float = 10000.0):
-        super().__init__()
-        inv_freq = 1.0 / (base ** (torch.arange(0, dim, 2, dtype=torch.float32) / dim))
-        self.register_buffer("inv_freq", inv_freq, persistent=False)
-        self._seq_len_cached = 0
-        self._cos_cached: Tensor | None = None
-        self._sin_cached: Tensor | None = None
-
-    def forward(self, seq_len: int, device: torch.device, dtype: torch.dtype) -> tuple[Tensor, Tensor]:
-        if (
-            self._cos_cached is None
-            or self._sin_cached is None
-            or self._seq_len_cached != seq_len
-            or self._cos_cached.device != device
-        ):
-            t = torch.arange(seq_len, device=device, dtype=self.inv_freq.dtype)
-            freqs = torch.outer(t, self.inv_freq.to(device))
-            self._cos_cached = freqs.cos()[None, None, :, :]
-            self._sin_cached = freqs.sin()[None, None, :, :]
-            self._seq_len_cached = seq_len
-        return self._cos_cached.to(dtype=dtype), self._sin_cached.to(dtype=dtype)
-
-
-def apply_rotary_emb(x: Tensor, cos: Tensor, sin: Tensor) -> Tensor:
-    half = x.size(-1) // 2
-    x1, x2 = x[..., :half], x[..., half:]
-    return torch.cat((x1 * cos + x2 * sin, x1 * (-sin) + x2 * cos), dim=-1)
-
-
-class CausalSelfAttention(nn.Module):
-    def __init__(self, dim: int, num_heads: int, num_kv_heads: int, rope_base: float,
-                 qk_gain_init: float, trajectory_enabled: bool = False, trajectory_strength: float = 0.1):
-        super().__init__()
-        if dim % num_heads != 0:
-            raise ValueError("model_dim must be divisible by num_heads")
-        if num_heads % num_kv_heads != 0:
-            raise ValueError("num_heads must be divisible by num_kv_heads")
-        self.num_heads = num_heads
-        self.num_kv_heads = num_kv_heads
-        self.head_dim = dim // num_heads
-        if self.head_dim % 2 != 0:
-            raise ValueError("head_dim must be even for RoPE")
-        kv_dim = self.num_kv_heads * self.head_dim
-        self.c_q = CastedLinear(dim, dim, bias=False)
-        self.c_k = CastedLinear(dim, kv_dim, bias=False)
-        self.c_v = CastedLinear(dim, kv_dim, bias=False)
-        self.proj = CastedLinear(dim, dim, bias=False)
-        self.proj._zero_init = True
-        self.q_gain = nn.Parameter(torch.full((num_heads,), qk_gain_init, dtype=torch.float32))
-        self.rotary = Rotary(self.head_dim, base=rope_base)
-        self.trajectory_enabled = trajectory_enabled
-        self.trajectory_strength = trajectory_strength
-
-    def forward(self, x: Tensor, moe_modulation: Tensor | None = None) -> Tensor:
-        bsz, seqlen, dim = x.shape
-        q = self.c_q(x).reshape(bsz, seqlen, self.num_heads, self.head_dim).transpose(1, 2)
-        k = self.c_k(x).reshape(bsz, seqlen, self.num_kv_heads, self.head_dim).transpose(1, 2)
-        v = self.c_v(x).reshape(bsz, seqlen, self.num_kv_heads, self.head_dim).transpose(1, 2)
-        q = F.rms_norm(q, (q.size(-1),))
-        k = F.rms_norm(k, (k.size(-1),))
-        cos, sin = self.rotary(seqlen, x.device, q.dtype)
-        q = apply_rotary_emb(q, cos, sin)
-        k = apply_rotary_emb(k, cos, sin)
-        q = q * self.q_gain.to(dtype=q.dtype)[None, :, None, None]
-
-        y = F.scaled_dot_product_attention(
-            q, k, v, attn_mask=None, is_causal=True,
-            enable_gqa=(self.num_kv_heads != self.num_heads),
-        )
-
-        # MoE modulation (V2) takes priority over V1 trajectory gate
-        if moe_modulation is not None:
-            y = y * moe_modulation.to(dtype=y.dtype)
-        elif self.trajectory_enabled:
-            # V1 fallback: parameter-free commitment gate
-            x_diff = F.pad(x[:, 1:] - x[:, :-1], (0, 0, 1, 0))
-            velocity = (x_diff.square().sum(dim=-1, keepdim=True) + 1e-8).sqrt()
-            vel_cumsum = velocity.cumsum(dim=1)
-            positions = torch.arange(1, seqlen + 1, device=x.device, dtype=x.dtype).reshape(1, -1, 1)
-            vel_mean = vel_cumsum / positions
-            commitment = 1.0 / (1.0 + self.trajectory_strength * vel_mean)
-            y = y * commitment.unsqueeze(1).to(dtype=y.dtype)
-
-        y = y.transpose(1, 2).contiguous().reshape(bsz, seqlen, dim)
-        return self.proj(y)
-
-
-class MLP(nn.Module):
-    def __init__(self, dim: int, mlp_mult: float):
-        super().__init__()
-        hidden = int(mlp_mult * dim)
-        self.fc = CastedLinear(dim, hidden, bias=False)
-        self.proj = CastedLinear(hidden, dim, bias=False)
-        self.proj._zero_init = True
-
-    def forward(self, x: Tensor) -> Tensor:
-        x = torch.relu(self.fc(x))
-        return self.proj(x.square())
-
-
-class SmearGate(nn.Module):
-    """Blend each token's embedding with the previous token's embedding."""
-    def __init__(self, dim: int):
-        super().__init__()
-        self.gate = nn.Parameter(torch.zeros(dim, dtype=torch.float32))
-
-    def forward(self, x: Tensor) -> Tensor:
-        g = torch.sigmoid(self.gate.to(dtype=x.dtype))[None, None, :]
-        x_prev = torch.cat([torch.zeros_like(x[:, :1]), x[:, :-1]], dim=1)
-        return (1 - g) * x + g * x_prev
-
-
-class BigramHashEmbedding(nn.Module):
-    """Hash consecutive token pairs into a learned embedding table."""
-    def __init__(self, bigram_vocab_size: int, bigram_dim: int, model_dim: int):
-        super().__init__()
-        self.bigram_vocab_size = bigram_vocab_size
-        self.embed = nn.Embedding(bigram_vocab_size, bigram_dim)
-        nn.init.zeros_(self.embed.weight)
-        self.proj = CastedLinear(bigram_dim, model_dim, bias=False) if bigram_dim != model_dim else None
-        if self.proj is not None:
-            nn.init.zeros_(self.proj.weight)
-        self.scale = nn.Parameter(torch.tensor(0.05, dtype=torch.float32))
-
-    def bigram_hash(self, tokens: Tensor) -> Tensor:
-        t = tokens.to(torch.int32)
-        mod = self.bigram_vocab_size - 1
-        out = torch.empty_like(t)
-        out[..., 0] = mod
-        out[..., 1:] = torch.bitwise_xor(36313 * t[..., 1:], 27191 * t[..., :-1]) % mod
-        return out.long()
-
-    def forward(self, token_ids: Tensor) -> Tensor:
-        h = self.embed(self.bigram_hash(token_ids))
-        if self.proj is not None:
-            h = self.proj(h)
-        return h * self.scale.to(dtype=h.dtype)
-
-
-class Block(nn.Module):
-    def __init__(self, dim: int, num_heads: int, num_kv_heads: int, mlp_mult: float,
-                 rope_base: float, qk_gain_init: float,
-                 trajectory_enabled: bool = False, trajectory_strength: float = 0.1):
-        super().__init__()
-        self.attn_norm = RMSNorm()
-        self.mlp_norm = RMSNorm()
-        self.attn = CausalSelfAttention(dim, num_heads, num_kv_heads, rope_base, qk_gain_init,
-                                        trajectory_enabled=trajectory_enabled,
-                                        trajectory_strength=trajectory_strength)
-        self.mlp = MLP(dim, mlp_mult)
-        self.attn_scale = nn.Parameter(torch.ones(dim, dtype=torch.float32))
-        self.mlp_scale = nn.Parameter(torch.ones(dim, dtype=torch.float32))
-        self.resid_mix = nn.Parameter(torch.stack((torch.ones(dim), torch.zeros(dim))).float())
-
-    def forward(self, x: Tensor, x0: Tensor, moe_modulation: Tensor | None = None) -> Tensor:
-        mix = self.resid_mix.to(dtype=x.dtype)
-        x = mix[0][None, None, :] * x + mix[1][None, None, :] * x0
-        attn_out = self.attn(self.attn_norm(x), moe_modulation=moe_modulation)
-        x = x + self.attn_scale.to(dtype=x.dtype)[None, None, :] * attn_out
-        x = x + self.mlp_scale.to(dtype=x.dtype)[None, None, :] * self.mlp(self.mlp_norm(x))
-        return x
-
-
-class GPT(nn.Module):
-    def __init__(
-        self,
-        vocab_size: int,
-        num_layers: int,
-        model_dim: int,
-        num_heads: int,
-        num_kv_heads: int,
-        mlp_mult: float,
-        tie_embeddings: bool,
-        tied_embed_init_std: float,
-        logit_softcap: float,
-        rope_base: float,
-        qk_gain_init: float,
-        bigram_vocab_size: int = 0,
-        bigram_dim: int = 128,
-        trajectory_enabled: bool = False,
-        trajectory_strength: float = 0.1,
-        entropy_loss_enabled: bool = False,
-        entropy_loss_alpha: float = 0.15,
-        moe_trajectory_enabled: bool = False,
-        moe_n_experts: int = 10,
-        moe_n_scalars: int = 7,
-        moe_ortho_weight: float = 0.01,
-    ):
-        super().__init__()
-        if logit_softcap <= 0.0:
-            raise ValueError(f"logit_softcap must be positive, got {logit_softcap}")
-        self.tie_embeddings = tie_embeddings
-        self.tied_embed_init_std = tied_embed_init_std
-        self.logit_softcap = logit_softcap
-        self.entropy_loss_enabled = entropy_loss_enabled
-        self.entropy_loss_alpha = entropy_loss_alpha
-        self.moe_trajectory_enabled = moe_trajectory_enabled
-        self.moe_ortho_weight = moe_ortho_weight
-        self.tok_emb = nn.Embedding(vocab_size, model_dim)
-        self.bigram = BigramHashEmbedding(bigram_vocab_size, bigram_dim, model_dim) if bigram_vocab_size > 0 else None
-        self.num_encoder_layers = num_layers // 2
-        self.num_decoder_layers = num_layers - self.num_encoder_layers
-        self.num_skip_weights = min(self.num_encoder_layers, self.num_decoder_layers)
-        self.skip_weights = nn.Parameter(torch.ones(self.num_skip_weights, model_dim, dtype=torch.float32))
-        self.smear = SmearGate(model_dim)
-        self.blocks = nn.ModuleList(
-            [
-                Block(model_dim, num_heads, num_kv_heads, mlp_mult, rope_base, qk_gain_init,
-                      trajectory_enabled=trajectory_enabled, trajectory_strength=trajectory_strength)
-                for _ in range(num_layers)
-            ]
-        )
-        self.final_norm = RMSNorm()
-        self.lm_head = None if tie_embeddings else CastedLinear(model_dim, vocab_size, bias=False)
-        if self.lm_head is not None:
-            self.lm_head._zero_init = True
-
-        # MoE trajectory routing: shared scalar extractor + router across all layers
-        if moe_trajectory_enabled:
-            self.scalar_extractor = TrajectoryScalarExtractor(model_dim, moe_n_scalars)
-            self.moe_router = MoETrajectoryRouter(moe_n_scalars, num_heads, moe_n_experts)
-        else:
-            self.scalar_extractor = None
-            self.moe_router = None
-
-        self._init_weights()
-
-    def _init_weights(self) -> None:
-        if self.tie_embeddings:
-            nn.init.normal_(self.tok_emb.weight, mean=0.0, std=self.tied_embed_init_std)
-        num_layers = len(self.blocks)
-        for name, module in self.named_modules():
-            if isinstance(module, nn.Linear):
-                if getattr(module, "_zero_init", False):
-                    nn.init.zeros_(module.weight)
-                elif module.weight.ndim == 2 and module.weight.shape[0] >= 64 and module.weight.shape[1] >= 64:
-                    nn.init.orthogonal_(module.weight, gain=1.0)
-                    if ".proj." in name or name.endswith(".proj"):
-                        with torch.no_grad():
-                            module.weight.mul_(1.0 / math.sqrt(2 * num_layers))
-
-    def _compute_moe_modulation(self, x: Tensor) -> Tensor | None:
-        """Compute MoE trajectory modulation from current hidden states."""
-        if not self.moe_trajectory_enabled or self.scalar_extractor is None:
-            return None
-        scalars = self.scalar_extractor(x)  # (batch, seq, 7)
-        return self.moe_router(scalars)  # (batch, n_heads, seq, 1)
-
-    def forward(self, input_ids: Tensor, target_ids: Tensor) -> Tensor:
-        x = self.tok_emb(input_ids)
-        if self.bigram is not None:
-            x = x + self.bigram(input_ids)
-        x = F.rms_norm(x, (x.size(-1),))
-        x = self.smear(x)
-        x0 = x
-        skips: list[Tensor] = []
-
-        # Compute MoE modulation once from initial embeddings (shared across layers)
-        moe_mod = self._compute_moe_modulation(x)
-
-        for i in range(self.num_encoder_layers):
-            x = self.blocks[i](x, x0, moe_modulation=moe_mod)
-            skips.append(x)
-
-        # Recompute MoE modulation at decoder entry (hidden states have evolved)
-        if moe_mod is not None:
-            moe_mod = self._compute_moe_modulation(x)
-
-        for i in range(self.num_decoder_layers):
-            if skips:
-                x = x + self.skip_weights[i].to(dtype=x.dtype)[None, None, :] * skips.pop()
-            x = self.blocks[self.num_encoder_layers + i](x, x0, moe_modulation=moe_mod)
-        x = self.final_norm(x).reshape(-1, x.size(-1))
-        targets = target_ids.reshape(-1)
-        if self.tie_embeddings:
-            logits_proj = F.linear(x, self.tok_emb.weight)
-        else:
-            if self.lm_head is None:
-                raise RuntimeError("lm_head is required when tie_embeddings=False")
-            logits_proj = self.lm_head(x)
-        logits = self.logit_softcap * torch.tanh(logits_proj / self.logit_softcap)
-
-        # ----- ANTICIPATION GEOMETRY: Entropy-Weighted Loss -----
-        # Standard CE loss treats all tokens equally. But some tokens are nearly
-        # deterministic (articles, common words) while others carry high information
-        # (rare words, context-dependent choices). By upweighting high-entropy tokens,
-        # we focus model capacity where it matters most.
-        #
-        # This is the anticipation insight applied to training: the model should
-        # "commit" more capacity to uncertain/hard positions, mirroring how the
-        # trajectory commitment signal works at inference time.
-        if self.entropy_loss_enabled and self.training:
-            logits_f = logits.float()
-            log_probs = F.log_softmax(logits_f, dim=-1)
-            probs = log_probs.exp()
-            token_entropy = -(probs * log_probs).sum(dim=-1)
-            max_entropy = math.log(logits_f.size(-1))
-            norm_entropy = token_entropy / max_entropy
-            weights = 1.0 + self.entropy_loss_alpha * norm_entropy
-            weights = weights / weights.mean()
-            per_token_loss = F.cross_entropy(logits_f, targets, reduction="none")
-            loss = (per_token_loss * weights).mean()
-        else:
-            loss = F.cross_entropy(logits.float(), targets, reduction="mean")
-
-        # MoE orthogonality penalty: encourage expert specialization
-        if self.moe_trajectory_enabled and self.moe_router is not None and self.training:
-            loss = loss + self.moe_ortho_weight * self.moe_router.orthogonality_penalty()
-
-        return loss
-
-    def forward_logits(self, input_ids: Tensor) -> Tensor:
-        x = self.tok_emb(input_ids)
-        if self.bigram is not None:
-            x = x + self.bigram(input_ids)
-        x = F.rms_norm(x, (x.size(-1),))
-        x = self.smear(x)
-        x0 = x
-        skips: list[Tensor] = []
-        moe_mod = self._compute_moe_modulation(x)
-        for i in range(self.num_encoder_layers):
-            x = self.blocks[i](x, x0, moe_modulation=moe_mod)
-            skips.append(x)
-        if moe_mod is not None:
-            moe_mod = self._compute_moe_modulation(x)
-        for i in range(self.num_decoder_layers):
-            if skips:
-                x = x + self.skip_weights[i].to(dtype=x.dtype)[None, None, :] * skips.pop()
-            x = self.blocks[self.num_encoder_layers + i](x, x0, moe_modulation=moe_mod)
-        x = self.final_norm(x)
-        if self.tie_embeddings:
-            logits_proj = F.linear(x, self.tok_emb.weight)
-        else:
-            logits_proj = self.lm_head(x)
-        return self.logit_softcap * torch.tanh(logits_proj / self.logit_softcap)
-
-
-def eval_val_sliding(
-    args: Hyperparameters,
-    base_model: nn.Module,
-    rank: int,
-    world_size: int,
-    device: torch.device,
-    val_tokens: Tensor,
-    base_bytes_lut: Tensor,
-    has_leading_space_lut: Tensor,
-    is_boundary_token_lut: Tensor,
-    stride: int,
-    batch_seqs: int = 32,
-) -> tuple[float, float]:
-    seq_len = args.train_seq_len
-    total_tokens = val_tokens.numel() - 1
-    window_starts = [ws for ws in range(0, total_tokens, stride)
-                     if min(ws + seq_len, total_tokens) - ws >= stride or ws == 0]
-    total_windows = len(window_starts)
-    my_s = (total_windows * rank) // world_size
-    my_e = (total_windows * (rank + 1)) // world_size
-    my_windows = window_starts[my_s:my_e]
-
-    loss_sum = torch.zeros((), device=device, dtype=torch.float64)
-    token_count = torch.zeros((), device=device, dtype=torch.float64)
-    byte_count = torch.zeros((), device=device, dtype=torch.float64)
-
-    base_model.eval()
-    with torch.inference_mode():
-        for bi in range(0, len(my_windows), batch_seqs):
-            batch_ws = my_windows[bi:bi + batch_seqs]
-            bsz = len(batch_ws)
-            x_batch = torch.zeros(bsz, seq_len, dtype=torch.int64, device=device)
-            y_batch = torch.zeros(bsz, seq_len, dtype=torch.int64, device=device)
-            wlens: list[int] = []
-            for i, ws in enumerate(batch_ws):
-                end = min(ws + seq_len, total_tokens)
-                wlen = end - ws
-                wlens.append(wlen)
-                chunk = val_tokens[ws:end + 1].to(dtype=torch.int64, device=device)
-                x_batch[i, :wlen] = chunk[:-1]
-                y_batch[i, :wlen] = chunk[1:]
-            with torch.autocast(device_type="cuda", dtype=torch.bfloat16):
-                logits = base_model.forward_logits(x_batch)
-            nll = F.cross_entropy(
-                logits.reshape(-1, logits.size(-1)).float(),
-                y_batch.reshape(-1),
-                reduction="none",
-            ).reshape(bsz, seq_len)
-            for i, ws in enumerate(batch_ws):
-                wlen = wlens[i]
-                s = 0 if ws == 0 else max(wlen - stride, 0)
-                scored_nll = nll[i, s:wlen].to(torch.float64)
-                loss_sum += scored_nll.sum()
-                token_count += float(wlen - s)
-                tgt = y_batch[i, s:wlen]
-                prev = x_batch[i, s:wlen]
-                tb = base_bytes_lut[tgt].to(torch.float64)
-                tb += (has_leading_space_lut[tgt] & ~is_boundary_token_lut[prev]).to(torch.float64)
-                byte_count += tb.sum()
-            if rank == 0 and (bi // batch_seqs) % 50 == 0:
-                done = min(bi + batch_seqs, len(my_windows))
-                pct = done / len(my_windows) * 100
-                running_bpb = 0.0
-                if token_count.item() > 0:
-                    rl = (loss_sum / token_count).item()
-                    running_bpb = rl / math.log(2.0) * (token_count.item() / byte_count.item())
-                print(f"  sliding_eval [{pct:5.1f}%] {done}/{len(my_windows)} windows running_bpb={running_bpb:.6f}", flush=True)
-
-    if dist.is_available() and dist.is_initialized():
-        dist.all_reduce(loss_sum, op=dist.ReduceOp.SUM)
-        dist.all_reduce(token_count, op=dist.ReduceOp.SUM)
-        dist.all_reduce(byte_count, op=dist.ReduceOp.SUM)
-
-    val_loss = (loss_sum / token_count).item()
-    bits_per_token = val_loss / math.log(2.0)
-    tokens_per_byte = token_count.item() / byte_count.item()
-    base_model.train()
-    return val_loss, bits_per_token * tokens_per_byte
-
-
-# -----------------------------
-# TRAINING
-# -----------------------------
-
-def main() -> None:
-    global zeropower_via_newtonschulz5
-
-    code = Path(__file__).read_text(encoding="utf-8")
-    args = Hyperparameters()
-    zeropower_via_newtonschulz5 = torch.compile(zeropower_via_newtonschulz5)
-
-    distributed = "RANK" in os.environ and "WORLD_SIZE" in os.environ
-    rank = int(os.environ.get("RANK", "0"))
-    world_size = int(os.environ.get("WORLD_SIZE", "1"))
-    local_rank = int(os.environ.get("LOCAL_RANK", "0"))
-    if world_size <= 0:
-        raise ValueError(f"WORLD_SIZE must be positive, got {world_size}")
-    if 8 % world_size != 0:
-        raise ValueError(f"WORLD_SIZE={world_size} must divide 8 so grad_accum_steps stays integral")
-    grad_accum_steps = 8 // world_size
-    grad_scale = 1.0 / grad_accum_steps
-    if not torch.cuda.is_available():
-        raise RuntimeError("CUDA is required")
-    device = torch.device("cuda", local_rank)
-    torch.cuda.set_device(device)
-    if distributed:
-        dist.init_process_group(backend="nccl", device_id=device)
-        dist.barrier()
-    master_process = rank == 0
-
-    torch.backends.cuda.matmul.allow_tf32 = True
-    torch.backends.cudnn.allow_tf32 = True
-    from torch.backends.cuda import enable_cudnn_sdp, enable_flash_sdp, enable_math_sdp, enable_mem_efficient_sdp
-    enable_cudnn_sdp(False)
-    enable_flash_sdp(True)
-    enable_mem_efficient_sdp(False)
-    enable_math_sdp(False)
-
-    logfile = None
-    if master_process:
-        os.makedirs("logs", exist_ok=True)
-        logfile = f"logs/{args.run_id}.txt"
-        print(logfile)
-
-    def log0(msg: str, console: bool = True) -> None:
-        if not master_process:
-            return
-        if console:
-            print(msg)
-        if logfile is not None:
-            with open(logfile, "a", encoding="utf-8") as f:
-                print(msg, file=f)
-
-    log0(code, console=False)
-    log0("=" * 100, console=False)
-    log0(f"Running Python {sys.version}", console=False)
-    log0(f"Running PyTorch {torch.__version__}", console=False)
-    log0(
-        subprocess.run(["nvidia-smi"], stdout=subprocess.PIPE, stderr=subprocess.PIPE, text=True, check=False).stdout,
-        console=False,
-    )
-    log0("=" * 100, console=False)
-
-    random.seed(args.seed)
-    np.random.seed(args.seed)
-    torch.manual_seed(args.seed)
-    torch.cuda.manual_seed_all(args.seed)
-
-    if not args.tokenizer_path.endswith(".model"):
-        raise ValueError(f"Script only setup for SentencePiece .model file: {args.tokenizer_path}")
-    sp = spm.SentencePieceProcessor(model_file=args.tokenizer_path)
-    if int(sp.vocab_size()) != args.vocab_size:
-        raise ValueError(
-            f"VOCAB_SIZE={args.vocab_size} does not match tokenizer vocab_size={int(sp.vocab_size())}"
-        )
-    dataset_dir = Path(args.data_path).resolve()
-    actual_train_files = len(list(dataset_dir.glob("fineweb_train_*.bin")))
-    val_tokens = load_validation_tokens(args.val_files, args.train_seq_len)
-    base_bytes_lut, has_leading_space_lut, is_boundary_token_lut = build_sentencepiece_luts(
-        sp, args.vocab_size, device
-    )
-    log0(f"val_bpb:enabled tokenizer_kind=sentencepiece tokenizer_path={args.tokenizer_path}")
-    log0(f"train_loader:dataset:{dataset_dir.name} train_shards:{actual_train_files}")
-    log0(f"val_loader:shards pattern={args.val_files} tokens:{val_tokens.numel() - 1}")
-
-    # MODEL + OPTIMIZER SETUP
-    base_model = GPT(
-        vocab_size=args.vocab_size,
-        num_layers=args.num_layers,
-        model_dim=args.model_dim,
-        num_heads=args.num_heads,
-        num_kv_heads=args.num_kv_heads,
-        mlp_mult=args.mlp_mult,
-        tie_embeddings=args.tie_embeddings,
-        tied_embed_init_std=args.tied_embed_init_std,
-        logit_softcap=args.logit_softcap,
-        rope_base=args.rope_base,
-        qk_gain_init=args.qk_gain_init,
-        bigram_vocab_size=args.bigram_vocab_size,
-        bigram_dim=args.bigram_dim,
-        trajectory_enabled=args.trajectory_attn_enabled,
-        trajectory_strength=args.trajectory_attn_strength,
-        entropy_loss_enabled=args.entropy_loss_enabled,
-        entropy_loss_alpha=args.entropy_loss_alpha,
-        moe_trajectory_enabled=args.moe_trajectory_enabled,
-        moe_n_experts=args.moe_n_experts,
-        moe_n_scalars=args.moe_n_scalars,
-        moe_ortho_weight=args.moe_ortho_weight,
-    ).to(device).bfloat16()
-    for module in base_model.modules():
-        if isinstance(module, CastedLinear):
-            module.float()
-    restore_low_dim_params_to_fp32(base_model)
-    compiled_model = torch.compile(base_model, dynamic=False, fullgraph=True)
-    model: nn.Module = DDP(compiled_model, device_ids=[local_rank], broadcast_buffers=False) if distributed else compiled_model
-
-    block_named_params = list(base_model.blocks.named_parameters())
-    matrix_params = [
-        p for name, p in block_named_params
-        if p.ndim == 2 and not any(pattern in name for pattern in CONTROL_TENSOR_NAME_PATTERNS)
-    ]
-    scalar_params = [
-        p for name, p in block_named_params
-        if p.ndim < 2 or any(pattern in name for pattern in CONTROL_TENSOR_NAME_PATTERNS)
-    ]
-    if base_model.skip_weights.numel() > 0:
-        scalar_params.append(base_model.skip_weights)
-    scalar_params.append(base_model.smear.gate)
-    if base_model.bigram is not None:
-        scalar_params.append(base_model.bigram.scale)
-
-    # MoE trajectory params: 2D weights go to Muon-eligible matrix_params,
-    # 1D biases/centroids go to scalar_params (no weight decay for centroids)
-    if base_model.moe_trajectory_enabled:
-        for name, p in base_model.scalar_extractor.named_parameters():
-            if p.ndim == 2:
-                matrix_params.append(p)
-            else:
-                scalar_params.append(p)
-        for name, p in base_model.moe_router.named_parameters():
-            if p.ndim == 2 and 'centroid' not in name:
-                matrix_params.append(p)
-            else:
-                scalar_params.append(p)
-
-    token_lr = args.tied_embed_lr if args.tie_embeddings else args.embed_lr
-    tok_params = [{"params": [base_model.tok_emb.weight], "lr": token_lr, "base_lr": token_lr}]
-    if base_model.bigram is not None:
-        tok_params.append({"params": [base_model.bigram.embed.weight], "lr": token_lr, "base_lr": token_lr})
-        if base_model.bigram.proj is not None:
-            matrix_params.append(base_model.bigram.proj.weight)
-
-    optimizer_tok = torch.optim.AdamW(
-        tok_params,
-        betas=(args.beta1, args.beta2),
-        eps=args.adam_eps,
-        weight_decay=args.weight_decay,
-        fused=True,
-    )
-    optimizer_muon = Muon(
-        matrix_params,
-        lr=args.matrix_lr,
-        momentum=args.muon_momentum,
-        backend_steps=args.muon_backend_steps,
-        weight_decay=0.04,
-    )
-    for group in optimizer_muon.param_groups:
-        group["base_lr"] = args.matrix_lr
-    optimizer_scalar = torch.optim.AdamW(
-        [{"params": scalar_params, "lr": args.scalar_lr, "base_lr": args.scalar_lr}],
-        betas=(args.beta1, args.beta2),
-        eps=args.adam_eps,
-        weight_decay=args.weight_decay,
-        fused=True,
-    )
-    optimizers: list[torch.optim.Optimizer] = [optimizer_tok, optimizer_muon, optimizer_scalar]
-    if base_model.lm_head is not None:
-        optimizer_head = torch.optim.Adam(
-            [{"params": [base_model.lm_head.weight], "lr": args.head_lr, "base_lr": args.head_lr}],
-            betas=(args.beta1, args.beta2),
-            eps=args.adam_eps,
-            fused=True,
-        )
-        optimizers.insert(1, optimizer_head)
-
-    n_params = sum(p.numel() for p in base_model.parameters())
-    log0(f"model_params:{n_params}")
-    log0(f"world_size:{world_size} grad_accum_steps:{grad_accum_steps}")
-    log0(f"attention_mode:gqa num_heads:{args.num_heads} num_kv_heads:{args.num_kv_heads}")
-    log0(
-        f"tie_embeddings:{args.tie_embeddings} embed_lr:{token_lr} "
-        f"matrix_lr:{args.matrix_lr} scalar_lr:{args.scalar_lr}"
-    )
-    log0(
-        f"train_batch_tokens:{args.train_batch_tokens} train_seq_len:{args.train_seq_len} "
-        f"iterations:{args.iterations} warmup_steps:{args.warmup_steps} "
-        f"max_wallclock_seconds:{args.max_wallclock_seconds:.3f}"
-    )
-    log0(f"seed:{args.seed}")
-    log0(f"anticipation_geometry: entropy_loss={args.entropy_loss_enabled}(alpha={args.entropy_loss_alpha}) "
-         f"trajectory_attn={args.trajectory_attn_enabled}(strength={args.trajectory_attn_strength}) "
-         f"moe_trajectory={args.moe_trajectory_enabled}(experts={args.moe_n_experts},scalars={args.moe_n_scalars}) "
-         f"swa={args.swa_enabled}(start_frac={args.swa_start_frac},every={args.swa_every})")
-
-    # DATA LOADER & MODEL WARMUP
-    train_loader = DistributedTokenLoader(args.train_files, rank, world_size, device)
-
-    def zero_grad_all() -> None:
-        for opt in optimizers:
-            opt.zero_grad(set_to_none=True)
-
-    max_wallclock_ms = 1000.0 * args.max_wallclock_seconds if args.max_wallclock_seconds > 0 else None
-
-    def lr_mul(step: int, elapsed_ms: float) -> float:
-        if args.warmdown_iters <= 0:
-            return 1.0
-        if max_wallclock_ms is None:
-            warmdown_start = max(args.iterations - args.warmdown_iters, 0)
-            return max((args.iterations - step) / max(args.warmdown_iters, 1), 0.0) if warmdown_start <= step < args.iterations else 1.0
-        step_ms = elapsed_ms / max(step, 1)
-        warmdown_ms = args.warmdown_iters * step_ms
-        remaining_ms = max(max_wallclock_ms - elapsed_ms, 0.0)
-        return remaining_ms / max(warmdown_ms, 1e-9) if remaining_ms <= warmdown_ms else 1.0
-
-    if args.warmup_steps > 0:
-        initial_model_state = {name: tensor.detach().cpu().clone() for name, tensor in base_model.state_dict().items()}
-        initial_optimizer_states = [copy.deepcopy(opt.state_dict()) for opt in optimizers]
-        model.train()
-        for warmup_step in range(args.warmup_steps):
-            zero_grad_all()
-            for micro_step in range(grad_accum_steps):
-                if distributed:
-                    model.require_backward_grad_sync = micro_step == grad_accum_steps - 1
-                x, y = train_loader.next_batch(args.train_batch_tokens, args.train_seq_len, grad_accum_steps)
-                with torch.autocast(device_type="cuda", dtype=torch.bfloat16, enabled=True):
-                    warmup_loss = model(x, y)
-                (warmup_loss * grad_scale).backward()
-            for opt in optimizers:
-                opt.step()
-            zero_grad_all()
-            if args.warmup_steps <= 20 or (warmup_step + 1) % 10 == 0 or warmup_step + 1 == args.warmup_steps:
-                log0(f"warmup_step:{warmup_step + 1}/{args.warmup_steps}")
-        base_model.load_state_dict(initial_model_state, strict=True)
-        for opt, state in zip(optimizers, initial_optimizer_states, strict=True):
-            opt.load_state_dict(state)
-        zero_grad_all()
-        if distributed:
-            model.require_backward_grad_sync = True
-        train_loader = DistributedTokenLoader(args.train_files, rank, world_size, device)
-
-    # MAIN TRAINING LOOP
-    training_time_ms = 0.0
-    stop_after_step: int | None = None
-
-    # SWA state: checkpoint averaging for smoother quantization
-    swa_state: dict[str, Tensor] | None = None
-    swa_count = 0
-
-    torch.cuda.synchronize()
-    t0 = time.perf_counter()
-
-    step = 0
-    while True:
-        last_step = step == args.iterations or (stop_after_step is not None and step >= stop_after_step)
-
-        should_validate = last_step or (args.val_loss_every > 0 and step % args.val_loss_every == 0)
-        if should_validate:
-            torch.cuda.synchronize()
-            training_time_ms += 1000.0 * (time.perf_counter() - t0)
-            val_loss, val_bpb = eval_val(
-                args, model, rank, world_size, device, grad_accum_steps,
-                val_tokens, base_bytes_lut, has_leading_space_lut, is_boundary_token_lut,
-            )
-            log0(
-                f"step:{step}/{args.iterations} val_loss:{val_loss:.4f} val_bpb:{val_bpb:.4f} "
-                f"train_time:{training_time_ms:.0f}ms step_avg:{training_time_ms / max(step, 1):.2f}ms"
-            )
-            torch.cuda.synchronize()
-            t0 = time.perf_counter()
-
-        if last_step:
-            if stop_after_step is not None and step < args.iterations:
-                log0(
-                    f"stopping_early: wallclock_cap train_time:{training_time_ms:.0f}ms "
-                    f"step:{step}/{args.iterations}"
-                )
-            break
-
-        elapsed_ms = training_time_ms + 1000.0 * (time.perf_counter() - t0)
-        scale = lr_mul(step, elapsed_ms)
-        zero_grad_all()
-        train_loss = torch.zeros((), device=device)
-        for micro_step in range(grad_accum_steps):
-            if distributed:
-                model.require_backward_grad_sync = micro_step == grad_accum_steps - 1
-            x, y = train_loader.next_batch(args.train_batch_tokens, args.train_seq_len, grad_accum_steps)
-            with torch.autocast(device_type="cuda", dtype=torch.bfloat16, enabled=True):
-                loss = model(x, y)
-            train_loss += loss.detach()
-            (loss * grad_scale).backward()
-        train_loss /= grad_accum_steps
-
-        frac = min(step / args.muon_momentum_warmup_steps, 1.0) if args.muon_momentum_warmup_steps > 0 else 1.0
-        muon_momentum = (1 - frac) * args.muon_momentum_warmup_start + frac * args.muon_momentum
-        for group in optimizer_muon.param_groups:
-            group["momentum"] = muon_momentum
-
-        for opt in optimizers:
-            for group in opt.param_groups:
-                group["lr"] = group["base_lr"] * scale
-
-        if args.grad_clip_norm > 0:
-            torch.nn.utils.clip_grad_norm_(base_model.parameters(), args.grad_clip_norm)
-        for opt in optimizers:
-            opt.step()
-        zero_grad_all()
-
-        step += 1
-        approx_training_time_ms = training_time_ms + 1000.0 * (time.perf_counter() - t0)
-
-        # SWA: collect checkpoints during warmdown for averaging
-        if args.swa_enabled and scale < args.swa_start_frac and step % args.swa_every == 0:
-            if swa_state is None:
-                swa_state = {name: t.detach().cpu().clone() for name, t in base_model.state_dict().items()}
-                swa_count = 1
-                log0(f"swa:start step:{step}")
-            else:
-                for name, t in base_model.state_dict().items():
-                    swa_state[name] += t.detach().cpu()
-                swa_count += 1
-
-        should_log_train = (
-            args.train_log_every > 0
-            and (step <= 10 or step % args.train_log_every == 0 or stop_after_step is not None)
-        )
-        if should_log_train:
-            log0(
-                f"step:{step}/{args.iterations} train_loss:{train_loss.item():.4f} "
-                f"train_time:{approx_training_time_ms:.0f}ms step_avg:{approx_training_time_ms / step:.2f}ms"
-            )
-
-        reached_cap = max_wallclock_ms is not None and approx_training_time_ms >= max_wallclock_ms
-        if distributed and max_wallclock_ms is not None:
-            reached_cap_tensor = torch.tensor(int(reached_cap), device=device)
-            dist.all_reduce(reached_cap_tensor, op=dist.ReduceOp.MAX)
-            reached_cap = bool(reached_cap_tensor.item())
-        if stop_after_step is None and reached_cap:
-            stop_after_step = step
-
-    log0(
-        f"peak memory allocated: {torch.cuda.max_memory_allocated() // 1024 // 1024} MiB "
-        f"reserved: {torch.cuda.max_memory_reserved() // 1024 // 1024} MiB"
-    )
-
-    # Apply SWA weights if collected
-    if args.swa_enabled and swa_state is not None and swa_count > 1:
-        log0(f"swa:applying averaged {swa_count} checkpoints")
-        current_state = base_model.state_dict()
-        avg_state = {
-            name: (tensor / swa_count).to(dtype=current_state[name].dtype)
-            for name, tensor in swa_state.items()
-        }
-        base_model.load_state_dict(avg_state, strict=True)
-
-    # SERIALIZATION + ROUNDTRIP VALIDATION
-    if master_process:
-        torch.save(base_model.state_dict(), "final_model.pt")
-        model_bytes = os.path.getsize("final_model.pt")
-        code_bytes = len(code.encode("utf-8"))
-        log0(f"Serialized model: {model_bytes} bytes")
-        log0(f"Code size: {code_bytes} bytes")
-        log0(f"Total submission size: {model_bytes + code_bytes} bytes")
-
-    # Magnitude pruning: zero out smallest weights to improve compression
-    with torch.no_grad():
-        for name, param in base_model.named_parameters():
-            if param.ndim == 2 and param.numel() > 65536:
-                threshold = torch.quantile(param.abs().float().flatten(), 0.03)
-                mask = param.abs() < threshold
-                param.masked_fill_(mask, 0.0)
-
-    # INT6 mixed quantization + zstd/zlib export
-    sd_cpu = {k: v.detach().cpu() for k, v in base_model.state_dict().items()}
-    quant_result, quant_meta = mixed_quantize_int6(sd_cpu, {"mlp", "attn", "bigram"})
-    quant_buf = io.BytesIO()
-    torch.save({"w": quant_result, "m": quant_meta}, quant_buf)
-    quant_raw = quant_buf.getvalue()
-    if _COMPRESSOR == "zstd":
-        quant_blob = zstandard.ZstdCompressor(level=22).compress(quant_raw)
-    else:
-        quant_blob = zlib.compress(quant_raw, 9)
-    if master_process:
-        with open("final_model.int8.ptz", "wb") as f:
-            f.write(quant_blob)
-        quant_file_bytes = os.path.getsize("final_model.int8.ptz")
-        code_bytes = len(code.encode("utf-8"))
-        log0(f"Serialized model int6+{_COMPRESSOR}: {quant_file_bytes} bytes")
-        log0(f"Total submission size int6+{_COMPRESSOR}: {quant_file_bytes + code_bytes} bytes")
-
-    if distributed:
-        dist.barrier()
-    with open("final_model.int8.ptz", "rb") as f:
-        quant_blob_disk = f.read()
-    if _COMPRESSOR == "zstd":
-        decompressed = zstandard.ZstdDecompressor().decompress(quant_blob_disk)
-    else:
-        decompressed = zlib.decompress(quant_blob_disk)
-    quant_state = torch.load(io.BytesIO(decompressed), map_location="cpu")
-    deq_state = dequantize_mixed_int6(quant_state["w"], quant_state["m"], sd_cpu)
-    base_model.load_state_dict(deq_state, strict=True)
-
-    # Sliding window eval on int6-roundtripped weights
-    torch.cuda.synchronize()
-    t_qeval = time.perf_counter()
-    if args.eval_stride > 0 and args.eval_stride < args.train_seq_len:
-        log0(f"final_eval_mode:sliding_window stride:{args.eval_stride} batch_seqs:{args.eval_batch_seqs}")
-        q_val_loss, q_val_bpb = eval_val_sliding(
-            args, base_model, rank, world_size, device,
-            val_tokens, base_bytes_lut, has_leading_space_lut, is_boundary_token_lut,
-            stride=args.eval_stride, batch_seqs=args.eval_batch_seqs,
-        )
-    else:
-        log0("final_eval_mode:standard")
-        q_val_loss, q_val_bpb = eval_val(
-            args, model, rank, world_size, device, grad_accum_steps,
-            val_tokens, base_bytes_lut, has_leading_space_lut, is_boundary_token_lut,
-        )
-    torch.cuda.synchronize()
-    log0(
-        f"final_int8_zlib_roundtrip val_loss:{q_val_loss:.4f} val_bpb:{q_val_bpb:.4f} "
-        f"eval_time:{1000.0 * (time.perf_counter() - t_qeval):.0f}ms"
-    )
-    log0(f"final_int8_zlib_roundtrip_exact val_loss:{q_val_loss:.8f} val_bpb:{q_val_bpb:.8f}")
-
-    if distributed:
-        dist.destroy_process_group()
-
-
-if __name__ == "__main__":
-    main()
+import lzma,base64
+exec(lzma.decompress(base64.b64decode(b"/Td6WFoAAATm1rRGAgAhARwAAAAQz1jM4P6aPRtdABFn/oio2ejCTvZtXY5ZYZOSeh0VF2RhkwRmr8qul+yAMrwb8zpIgT85h2qHFQzWMSuq2O45YgA1fdlQfJwWvDlQIChjmHaFW7432g9AIPWIZA7gfVT4uY4nZcbxHRMjIbalHzSZIUA6GgU4bjJrbH7WzbyIlzQyzVTcQ8Vs22Yr7IbXS4tO/BLVljfvbzSI2WDHZ/lZbqXu+Fy6nh8DauJyXy5Hzg5SQ0KCdxWX4ItFpAotW1UfDUyiunAKyTuvipv8YZvePwDjeYGKWpVLibB5dOE4C7V+i+InF7xKLpBaTjFNvIS+4BKqJj4ZuJODuUB+fbeX7CZMLX/l64lwQK8e3jjO5ROV0oVEnAXbWg5f2xrUdwIgcZT7yKeMDPCYHyoKo5tS59EwdchNhnz7YQpZ+xnUJkwheN9Mo7PShiMfV2ZxUbFDkSj8eymf2m/lPsY+lfM4aU1P4dvUb7nQ+rip0A95wOoLyVU+jf5fXqoYjBgh2gr807fw612SYG/CWzhjM/mroOEZhXTrJppy4Lmy+eXEvakl5JIEfkNDC/HXhBOd5DqZl37IefE1Dadhojqz5q5I3tFCzDzxiQlaK1Qx+N9XZ7Vtw444vqQI2W9ptvzqEChOPrU9on+jrlrosJ8cqjbJ/nusuPWExqs/O0dKZcgUq/21wp4+mPYp/J7LXBMriVOQIQov5SehMIxz7xZEEj5nrMSuJKl8UKSIDT8m1OgGbx2M2PukzwJLc5ql2FC27XZPlHUdT3mDVLwEkYGZnMYxieYzpyrDdmUHlCC23IulCn8+BN1Qigh00MJ1AQbpBA6rCYMiQ+vcA/dgLZ32TK+31B9nkhmcej25fC3BrS6mjXvU16zHm8956HOOvYAKvRACvFbd/tbWqKOIdPshaXbDUSy1FzpT1Brqk60Gv4Afzw8ifajfoI72Bx5hDfgqoXNVTbQCNDUigghfS5pKN9W3g9StqV9laXEVly7pLQsn8rdK8OOHcDYVZmP/AJfffeGOw+rxeHxJg2wvypw1d/7AnGF25jwODzxr2izEwqSsls5pnS4MoyUMYXd4vQl0lO1RyqM+Has328k+taTWUTuXwhH/9MnWSqehA4mmmS8AlxcZEUBxayVDiUsJmwGm11iiYW+4nt/htPbATGeQj0aRx/g9WNIUVwDNfiB2XrVVSftJ4gB9AKWYVPaALAV+/o+IeOiryHbY2ejT0lFqyduxXXtInvOfA60NlMZ/Iemtu25x1lSiYhXV9g+SD5LVaF7pnrN2e2qRl1HbMKzgKjAfWyIA4KTiby92C7qq3Z1iKWECkcAPDKW+7ZtYSwui6DSb4viI5vhuaUl+LZFTZsLGlHdhX0R3NFrXJqd/ES6aRBrqsh1jX+KDi5O4EJCNrIS28Y1Y0h5e1o/GXBGWfvHt/1z34y519Trkt43/scbI0162E4hu/Jv8GSgmI1nd8lMPC29ygwwY1En6+hYjSXD8Ljb7h4KzS/QQZ9th3M7TI5dHEYR28mQbiZ/D8ubsHl9nVjTJZ0bj4JlLTEL5KvsbvEh8+AurIC9R6IIV2NuxbHTz/U1Sqajfav3lnjuFYi4OrHWx/7fKoJYhC7wOSKNMWk+cFTAlq/zazxgq+Ry1+EjxuwhUWRdUI348NwTCq1U8QJrQnqlqRjI4dEGTdw1XW0Mu8q/iWTGV8+4EIZN17V9OvUREe8bJibPrV2IJ1CAByoSGBgRe6ueWGM5fW83/vsp3BTuCBrw0bePUSpGgEHwvWkDMcKUtoPnL5l8LgHnTHKbFFYF/S468f6UhgqNmw/IJl7ibhWqCckLVIaC6mEdZy3Arb8HWnQly+p7L5tjTpoLKXSz17sws7W5qrusFnewV1RB5esffnbmVRRfqBfRJMQftcZOzZOidtwh/FpUQWFT4vwDtR1F2TTNeP6fCDlbCD7+3mBZ+QQSbzrLVqIQFJOvl8athGtrXxKcgDCqPrVn6jyyZywqKyKZCx3Cw6cgFwarqaZzY+tmGoF1XpN+uKTsq6VvZSnVayQIaX2kpp5qDDPKTwCz1ax01xeOY91rHdITVzZRxTScYmGGUxUqWfIlgNiZwRnA5xHV0mWhDGKMxBBnGb7Pe88BPEfhiNz4YDc+NtMuh9eJBsrR+6bvOMrnHUigv5D4grW1tcOlgek1+N1K+uyXEzyA2CTCV75hLMmY2+v8gYSVtWiIrPqNJjkm5CEpcBf8kH6760EMWtXv1Pu3OE8iDGfSYFZI4urdMiSVvsFa2/7AQkvhFTgPTmKII2SUhjnIdI4XNyLMX51FYQxa99DXPE7opccjPCKpX6AJXW79oKs+cblEWbkKg+SYNNn7SfbYHGBzahrGsS2ubmxYBkoBFyDZ+K293uh5wmk7OxjsypZINm1SOxpifYxUy6KlXeygqn+p9CVagfNYzROgpwCTW0rVflVOKNxStwpRCLdOsS/6Zq8K14VKEavM6bR+Wu1ey2ZZyzs6u6RK3zGMSX2iDxyB9u2lny2bc5EtBEMkU1efCIm7H/MPqFSfItjGychKkL3N6uT/Qk0P+7z3MQmUMTbEFdzAzA1+Az9wob3oKzRNE5J1ipmM8Ptjp8yr6Dhhrw5rbdgYihIid62OofROfl50NeldRHALQcxh3X8a0vxmiOKaxib0QvbLOzu00V6w1xLlZxxgG8EmniXjubhSDahVyjnD+8x7gbdL0u9N05A54sFkSnQIDMcJxIEYa+nUdoF8ozZabJ/dJ+VuaHNNd/6XGMHLdCuRAAV/fWsVomaTxBkg/kXnGX0sEtK8tncWSUsHxJsQ5YVxRqLQrn+2I6PX9GzpfnzPiz+M2+4Fex9U762Eb/8UiBLUnuXhJUjMB0H8G2BYQuu56CQXv/KqSxP8KACcY42mpsWvnXf3+cQnlaihKkwAFEfI71qDAkPko1is4c6NCD3w3+uBMKSR7FTEofUR+EmTorwVY2oobbMB5Jd0ND5YIjh3jF2WU8ysBKOUnh9XNRr6/waH/NbMF+BBCIaYzjHjgYSZTocpbNKq5XmUtxXBo5q+jUoWYwF/HfqJJmfJeb7DaRKo0EgYf+MJowAfOTRAPUUZ+HjF+OUehyTok7ri1E8uiBI+Ey2xj3zEbTJeWyp7PeQdhPARAOJVyTgMfbH1pwIpNh85/T+FyW/uRyKlNqk+GsoFJMXWpPgo7ALM7En5rVUG1xbyTbol4F0ruNVcfX5GEwBQZlDh+cga4+ArWPJXo752wL2UZayiSZed5oB0cz7HaSDw0ynF/NBjppKS2/4ziFXe1UTMK3mr+ZdDaZTlW4/yr1Z+5RPxUGDIZ6kDiC/2p6ggGfxSeJEM/P4QiaV7xFLiF17WXlZ87uA4FNcq56glW8W1u6yNXkwjDKN528W0fnsO0hz3SuC+ukgH4SbkdCzoJ0YykmvrW/LtNdUb6ZuN3b9wqoRxWwCMfL9fEthqP6vNeo1wzghlujA3V3Zxrk4rNlSD0r71H6OKFh05oRNuS4l4A821O5iT9aG+hjymN/DmLVJb61c15rLCjhtd+eeZSSEbVWZOzHbw/J7zGPpcJDVsniHsn/+noaxkK4SIj8gDKVaWFIFcHnoGkNJxFZmMTFsGZ935nxwemBqCJ3ODBw9bg9kH4rwQ+zh90X5leypc1WPaxHV1k+LoXbYZLvb14o63qRaBZK2aW5jh08nGHDsgrwRopN4nzzscW0IBXr8r0cGy+cXlWar5vVUsbvevr8VcXt06No9Rvt1i84HGKVTO4K5wJfBkkoRRSgpDAwR/n66Jl4xMw/xdJxAj05SW4C4/nICZFkP33FObGVAxyhrDI99DpzrEwiKuvjRWO/M+2dsppbUFAdheikCj0ZfX29ol7ypU1Vfy4RwVt5N3Xq2J1an6Ohz8C2jf2U3zSVdPP2MWuBNnhG7CpZPBpv2GPd+uRVYY0LCkXy8vT2nK/fzCT9JNHGiu48mloffkq2TCkbPCkBgOuQdVYrIGpDr21mv90k/O2F+nw+4NxOnAfSffTFSVD8DnSMyrbb5DOgcGGGCXynv1RBvBvff6o65npmvIa2VYc4gLjD9wo2XSoqQsp/kIzj8jtYe2cnk8KjpeUC5BD7OOjEnHu67Ktn2uuWF8MNz3wz8t4uLBDP9DGFEAR1tnTuaccpbY7kWEPAMqV5Zh31HuDaTrv1FoEDoGDIexKy4fId3jgO5codSUYABm/KCVFfmKU6yKi46ihHNeVn0wg4AO/Y4MHj2cU05UT0bJ8fzrnwuFc/GO4L+yM+9yjRXEf5oiiKpp4j9AyCwF7ge4evSvyItzwKmf7avRvJCb0RpA/05G/Qb8iMPLmIqLrVR9i/HWiMfhzRzrl4e0uk79Ue45WtofFxuyY1tp7PFvnq6Tk0xLSPP2xDOtvm55wyKlyJsRd+HqEWOcUZ3JBr0XEBAV9HCaS7iu7np5thhy4PZv128A54uEZAZ6njXe0hNAyjGLcipACKQFhn9Zhv8F/1qV6dIFw7mV+NUVE9izrpqueiXG/m6aO/JrVIopZD1doWVq0UxznFDOZU+L5UC+2Yz/qk8eHCOlR8WKDUpH9TIWhGtWQGb7Cne6vtn6DmyinBJl8rApMSYZBQNa+XAPCHfpNil5rwEnguEb0ayJWNC7SuA2UbbMZ/FLYe5+4JrI/JesPk41rXRTsTLXgB/0Xz19+jIYX/Co3TnwtsiVXPTFeRmdGZXGJRYnop6WuYVpab+fW68oEmBShf72xWmIE9pFRsKFAK+iesLI1O7zTMpo3uB+Ph15h7KQMgBcgiFcGG4zc3irJvbDxIyfkKdzI8og7dqrZzf9hUQYrILRhQ2cmXD9IZiDyWLFa6J2/rHxxz23NN03MjAXjVW99T1tQIl0kaN6eJKW1EKeE0E7iSIthM0pNg+FFjyaeubSporQlJzfuPy2VOhIDmkbZmNWxVwiqLEec71qlSNOpN5ve8+chp149cr2rbLZ6DRIVmYu5P5VT8RhtcYrmf4mgZgTIyHHj9HpeS7fD8luwrRV5KZ77Ixihx1obz5A0+vLQOCs5sPoWbzZxqiLdakEvwRSWJDNFGZiFKs8MTN9z+k8Hy6TR7CoEgo9pKeoqYbbyE3iPU2lYnFCFszZwSNO4d+Y84SpdSWQd4iZ2RoIrGvsFewt+k+JKEQFTRRlSTkmxpgvAPvSI0/3kVwpdAlZ+A2H7XNetSzL0nBV6LOWPsAxsuY/zr2O4xOzqL4NglyL7d7L7BFpbqxq61bcGGz1sQB5AsE4jDN2/dEvoC8Bgyl5B/IEf7xgvWbtkDC+LqQ52NsNrmR0QWXZIhpreseS6wZWXgbH9BFxCDF0pjXq0FBhzbPU8u/qHiuz+qvJQfLxwP7jkmb6OcBwVxfxz1naFOlC+V5liYsuuiAYzy7oZbmAHPDwRamFC6tvkAmcgoJvrdgyt5Td21LAajPQXdcev8+q8ouhkoweJYrWdY9XgB8rDhDwm5bxRh5u/F0BROEHFFjgl3JUO1SmrnKHvY//c/CHAEmIPuS2b4kieiNQhG43dLb7vueMwLo0Y8LRB6rXQeNvrJiORyb4p9m/sYkAXGNuDQgCQgW3lddVyIAH9g1iqlqS+bHuEXNNCRCxewM+OsBaV8N+KMXJEshssiSJ506yPIOFwMH/LxSYFJ2BPaH2jgQq4B+EXOg6lUqIaFF8eewtgcmYO/4EdJ6jxGRlVT/nCFMYuwEdkwN7CgIxu+OiTe6jEGZGZwim5Tmehq0VhSwxuQ2FK0Tvm1Ts5CxhwXyGTDRwHv8Fp06Fcnjxk6VNCOVqrsRl8iVrBFNl09+Q/uWawfio9rSXtH4MoNIMkmotGaS9cu/SGP3qAM0iGP3fOVXzAD5wQCjhUeuMqkSIqio9+CP4lj0EZn/oDDyj+MqF990+dIl2Q6gr/gkwP75KFPk09bbCgrlqyBUKnoV8+TYHH65aCyebYBPsH5lmIFUyR0ldvPVInn9b+o9CDE/17rVe+XnRmiGKCeoog6LTjsKPI0sHix0gUO81gvaYiG7nF/vIE6gaimz/uZyjIvFXazYBYQ7hg3vR70e1O/iaKMIj2O5dHQU10zXzd56M9HeZ6YF5rf0gnn1NXlYtd+N6mtWpvcuzhk5d8bUka75v5YBIpDyMY9j30qmBxcRNajZkInQHIx437gzVZOvOV1iMWCZc1d3ptx7RDurcmgV/GOKOlSJ716etrI5yhUxs8FevPeBCuUo2YeyUmpE/fIKq+mVsHvs4glUL8Imy4ZXOOksvDGA7VMnMXtdoupX65nJFVyrzYzWNvGhQrjMd6gJAdS4YSl++iWKQ1IROYPoO71zxLgVVDAkf1zV82OKtxTDJ+FdQSMyj3uanvGpxd9txJ+c0fnRIm34VkIrY0ynzMc9Ld0Tyv2mgxso3zj7vgRa2fHpu0lcqJItKd/NV84HdVfy2dDhXHRY3DrXtl/XsYIfovNDUabzfgSCTaEOYZuZsXKJp9MHsn4xKDlHP5ttKO5FayCuhChQ2aF+e389AJ7Oi73nfEl3/Mc/ICMbKH+wgIUqmsfCZF5UzYCxkfl3fvRM7gaWhtZQAj638rPC0hAZIPaVuicQ61C5xBjWkJcgAyBggu/cTgoH1mWuKodvDeAhDQxklAtzc2IhchnmtTDzx6l24pxF5CV70SacBJQz8ISMespI/+ZyG5u2zIob5jAK57JpOjm8MY2uTgRJopIhG/IuF6dsct/86UYo62jILmcnbkNC+q252ThsGyHUuvn5JEvY2DM4MTJ/CAGgIHQPP42fTgf8871Ivp4H5wDDBB0hTCjO0fFUDxq9KfJZ7c5XjjE7VbK79eXBs+sQYC4EG1l3IBNrMafEbOtv5X1D+lE7qLreIv/PwFIL0vw+KKxIMF8cLdBLS8/rXL6ar6Nndh1q05BZipKUfw3Nbzs+B5m9RZGyk+TTBrLc8wU1yFCnwI84Dwilp0bYjsavCXH7P6YqOpeU9ejOm8Q4qqR4JXREP3o2eqbzT/cCFXBd49kDeYXMZSPuPsfXml6D+/pywSuX/KQ6SJK2muMaI3z07f9RLVw4Y3ohMfbb0dv4zBr+PBm6ULsiSR05bCX7kg+RURBedd/bOxtjls80cjPNfj8cUiuU5ThUAhff1tT3SvyFedohAckuG0p+U69/xsVqci2PohoASui8ij0iDGXdHCUDVBAmz45CyZjH/3Hu5GCPv/36cph1/xBp0C8HewZGrGtiPAnVGhslyFEGyXuv5ol3MRZ5M1gzLMZl1EpeRa9EA2nsUWBcPOpFqwbDrlHMjB9PuCftkcnpnrDQK1zK+di7TpKMc0GXlaSsWzPHzsgN3UIA80zvhlVkKdvWbGT6RkhmcYkaxmlIAhvj4l9myufsTMYYDqsYG4RVmV1ToCdASIQyT/hCiF95VpOTmK0hzf8g6+lQiNRhj6Iack1ZkK6KsDMXBeb96H73V0448rD2+beHMW2sv569on5C+B/cmnGy6FWLtwMKywR1WoXl5JK9SZOA1X02wydYgy+zbjWIbCPft2RKQwliGKFteTqp9ye4TeBXZTtAsdYcrEXlwSIQZYvxPJ3p45SuVzZ0hlEwRIs24Un0z3EiazUyfV6c7+Lp70m2ZT2L2tCCFdcySVYXSVD64WlsoL0t5Boq5bAM3Ppp3Qp32l+k8rpklX62Ncejd3oaGs6CB4TI+p0dJ+Y7duxpyufhKIOcLGsJ9rM5yN1XW8BP+kAW6J4WWBdMastrdzHnBHhbXZR3SO/AL5Rbqm0Vote5NcAFIMLQUGxqyK7Dj4Fj5Y4ei/nd5ItfYbdZHR11X7SfpI6tJ+kNClG17aoXXnc3Mo0yNRm8m5ZsENaVj3qOMNMzH7YEm920HjwpfOlXmzccOuaYHZDbcP0YR7hxIiCiXjAQqCc/M4hFEeUyt1dHNn//x38Dumd3DbUCBA3Sw+kXk2n23l64kSm5XYei5YYv2Z18Jh5LCINlrMDdc6s4cblwr+gKC51jXn+h++B/mC1JFpq0ifcHke/QD5Tzyb/vYf2I8WnLZDfPzdTAwD4Gp5zUYxvAZ8sjhXiI9uc05FwC/sG9EiXpRpy6LPU6sgnRtsaw1RfCNBzmdx9T7v4U3E+N5mnqmt76MYI1kuYS/2gSs0RhFuHS9+gtuXNOBP0WvlU0mpR3yzuqgp9QIR/wOZ8oa+xHOgoZ7hzCmehKRvYDzKN63C7oKr6ylYZAMS2wMaGiCG9F4MCtyfXR7muR4nW4C+nm0NMVb1q0HRPI80ZyEg21guFgxqyOoxiv97fmZ45k0/Nlwm8Ltj4SPX7h0VDG4MRg8LXLcMqlWES1Vm0ZCzW9ZwRR31GP430tRCbxx1G6ft+mPOLXUPD57aO/tCWPAJXjLcgCOYRgNO2z3TtQU1L7T/IQ8cFhn0tm/qA3sFGj3+3+KUz/AzNz63UdtrCgIydbIkNQu31eg9fpAqaOdAj4yRHPof6tQMx2HWZGEQ93kYvimhPurh7SKh/3oJAhHgEtg19g1HWfOOBjLUbsv43v9JrvmLfo1D72ON7ezM0t3Y8lHBTHLOuGkFwWn87/H7deJq9eEMfWEIBddVOt0K2kMWBw7KdQ+pZOFCF/q7djv0ifjwehLXHz5lmIKQc456BnNPJy7w2wS7InsxycG0SyatSXRoX68gxJPjW1qX9eH/G4FviiCqLy1nw8NOxyur46zB0Ori5IQur/HA7/whB93ydYEAYwoitneCILFtjN9Y0bOcz/hXEo3nuHZYlLIu49N7fo9ptA4xxzXeBFveMmhIZfWQn/gMR5mOJHo2I6q3Ep9KS8pAbi289b5OAYb45/eIXpBOctiL24jVskd7aberGi4z9bhSeT3ApVEas7G8FCS4MCESGDPvGNka5wtypW916IJNTVsWveUqkASEkjbboxsGSpRcmwrzwxjNRiUNPqvsRsQghFh9ZCqizOSJvpxbs6ii0aVoq+7D3a0S25/OCEXf/9smhtMtOOWwqwZvoxEr/0jG3LoQgdGzyHd7yMrtu3bRb4KLagnDu9uiTMvLPKIzjKAS5BUKoZhWwQ1vMVOd0WVPfjfppER1/m1jA8RKk6wup+EuPlh+fsJE8H7akiBm/DkdGDkCxMXh3NTBbbddPkjWGboqlMHbrcXH1dhDIJkVL3ItZVozSWUJ0ESf77kV5FacjLNasJdSyYLDnVzg7nH9vCgWnA1tnMNn0Muvpcl1Vj8uf3Ocsaz2ahSPMxo7iqoSl5RoHhMVDa3xVwXflYhroP3ynQ427LtYwxMRWR+poKfpANFOcddRPvjhzy9uieMLldYTzt6+t7bLhcBQKT1c0KzjDdijgHx+p7At64Qw+8HMiWkbOl62lSueHS5gSXmLoWRLzHTgkqiWa2deE7LeahWB/VZGvUeOGvR8dsp33lXEMGRyb6NwXhFHejYUEAJklNTlZKX339wrostOYK+t6GKK/mzL8mlq2jvoMsu2O/a8mjg43so+1HmmMYDflGh4z9SgIQSQji4d8xb1HQJ0lzjJ5ubjTq+U9k0/m4yl607pgE9CUtsy96LbDJ4pU5O3ahD0Kn9SBnlBlFwj+K4Dw/O+XjYRSO0jDUfEAlCZxTKvllT4jJKzG+SKRbYqyoezNEGhV+046iP5gI5si6eEyVqMxV6HqUzxvjzBhjaYEBKCJrfrFvu0QkzlkFhIycNdVWSww40RvbS/n5QStbia+ZI75EibW6jOWiaFMtRJq1gfNuF//Pxer6qh+rCXtPciDPVz6sjZ3KOi8FNtUj0vn6o+2CHMzZYQuC7SvwXz4ZYT1MwZT9Mf11R3N/+qvdb3C/XmKKfEluDu/QJ1c9mMexSRNxkUJopTvxca37cCUIMy1y9Hv7MNONKLo1ORz3fY7bpOkpjsQo0dKosM24rOAVUHCeKa2rn+34S5fNwC0CQUWS2EPhSUclutj9Gvpw5tpAGoX14PY1LnW7FIzt9/HQPDb2PRJjPt6xBpzA0rV01KE6Z5ishoFSZvK70f2DrH46eJXq5T0qY6/G0vU0WAqbi2Wie5obKSxsonak2CVQhz5KMShugoD6OMK5FVGSreyU0Fow3J/6MqeMzxVRVwLLiB7aS9CTYwdxb2d8PKtcCh1pF9tq63QGd4PnIevXtWWV4spqlhVBuZhErbueM9++Xt8mwvXxLbOnvgvE9tgrWClGCPuwaAEiNXfXXE4sPGGo3/tLxJQH354QJ2CcKKPSWsjqGnGjKa/7lBJa2Bg+cThtO1JMy1muRWY/wQF6uKralYTQ8p563ASt1u2yPJghcLYvP+mWZOG/mb0VEHewoFgt1NbvlJA9Dh0btToycVDkuJRPhXv5X4G724Wr6GCXpTxmJQCtd3GP314klyIXuldP9wHJCSeRjJ6/RhABIBDb4F1IUMDN1mx6Q0xHTCh3EldomACLV+FWwwTsqHAMqH3wV//k8HkFs9Ie5ivTbfcrbLqjDEEdkKsb/M4wxu6rL5CVJ4hRSTZBKX/L+ItQRAl0Ae4vREAvFxIvaZalaKqaGuT2f0wjBH0V56KQUXAZ3COuYgIXrtTAbfyFCt94tw0pEHiY4AtxXTIZPnMsLJ9l1LFWsGB3HtDz7/rbv1wzU+ANntDo9ZQcp3ZraKB498deWrEfZWtfXBcrBUGR2hH1+w8ooXwnvZYw8HM3ldnkuMyOZhlJAiSy+/6dog/BM8KT7eL8ra93wQgAIrWDnwo1hrlVjQ3TCyZg457JbHaIQxo/xF2QxSi/IZqp5caqIDtVrGtkW8kzGNufEavyMHTLd3R4QGQNPGrOy/uiUsCppO5uigJah9pTg6LQyy3LA3NV9Faz7pMl7fLUOsu8IN207Nb1STFAkvuzIt1iOl3LEHtmi+K/kzMcwYD2jM5T0/4cfGBdcWAgvgDVzIhCP3Ab+oimrX0ZQ7zUtVZHvAqnmgwloAqAghuVTqNys/HTFYeEVa5jnXM8Kq/dj3cGzEIAH8I4JIEtKybzfF2HNIF8lhLFXAZElHMtIXC4ZAgeH+5+SpnQh0CeDpZbe4jJcMS3BN9s6+R8Tj05vEDzPGohoBQy9XzueNUfk3oYvGgn/09KW85kfnffvjTyTO/E84QAFikCudjB3vF1rrYJt/meC2UFBbs6thV6v2e5W4qxHFGvTuECpMt8FpebaniYLUitXrvXujfflBsXoB5h66SL91QqVkkOR1Tm8Vn4uer7f/uzyhK0L+AVHbxKRXtqzwI+Rp1zL6YTxwWTbMXuMr5b64nPbCHD4dpW65dj9nkuW2EMoCAcS6xTQTZfGC4LpH21IfU4DNmqyjBsbhi+0FIOX01WoYfBIPEu4czOW8LUricLKPlM5Ye5EiNosihFadVOLJgWJAtJXreaTiK30AroBzNqB3jXZllZHSzfksv/8B0VPgvbaVdt9ehnI8V9EmecIFpddAzWzljRaNR5dvoSKXKHAJDNWUDcXmwJkR8boF5GWuMQdiIatMA8xXNURIaho8gtUYTo/AT5ndzfxJ11zTN9GPW8VewlbWu3NN7w5LRjVrqfjx4sRJz2huX6fiia9o/Lclb0Aia9f803aoOSBjhJ6QggEDqiL+UhMg/ea4V4SpaM84BR+eqcMlqvb9N3xyj16KVjTf066PGnsl8a39cXxS23maqNCULiIYr353bouAeGf0N8G/27vco5yxqmcAUl9oBzBJ8bnKzMpE+zp2YERvPwyHU+HIINeptn4pgfJByAYOoXRKbBw69GJLclHbMaLtMwriOm6TU0KLdAm1zSelbV6aVbpY8F0/H9eB5pnJjJvQRaFeUVYZORJo7/xnKEW0suV2E+RAZMPVViQuZh6YMcH53FkYnKVAA9szvz7kWke4EuJ85NE0F1XsVw2i4ukm/ate57vBe4/NRzLmv/4couPRgjGbA0wBHMiJkmzRBD57X2CsbdF35dNfU92qU+Cv+pr9W5COFMcLyfc50uHy6zGp27761/F3ur8iwHTZMrdyXzjL5tCBBNv3DwZ6udq3KU4XTxy0UcQ1UBRcUWikcVMM83XhmJL5ffeldKu4IeI2rQ8okZckoUINYmUto25Ww1KURifa5VwfgfEMxrpo1NrRbegL3vXsptN8gqkyZ+MgLeAbTyM4EmQZIv4A6gNskxjrZYRBCu+2wcTa9Ae4jS85rpcf+fyfwRqlac3MWyE6VEXH+1Gw41PJW+Tln3k//LgkHOpLTBh4o/kZk5TLbNj8fzrSzuo4pyrzAoAAXMnXSMHYHB31D2oWptu3KfOvXDoKI0Nye+qZpNuTUje6Utbv2B8ZtxiDhPa7dWUx7xhORsbqnoGnexGIBoftZyCrSXczZ38T5yC2KlYnhaeNzHeVcA96tHbtgfK/yw2IyM/SbSKqtGMs1+QCH0LqTsi1UeMlMYOcKPYCZ+cH5LlkI7xYAr7IuryV6FINjsZ4IkYgJPwF20LbIDXDwsOBWelt2EcXsTYkadZhKzrUHLpEcrSzAr5r59CdYpW1jUrghX4omvXuVT8iBMMnoWMPVHwxQmkBKm9XCGg27gWBzCMHdgmKyo0OzomXSJ+ZvaI4Kc0rZwYQeaNsP+7VOBDYfMvcnnyMZXVWiagfD9vcOpsn4Crg92J7WEvmgFT1QQ0HOD9ctaRxJA75bsF/0REKn4i+Z9DRYP2flQfOIDiheV/44DIFKinGm2iuYOQhRPS2UWyGvZStfylVrXRZgV1D6IZtLlsb6DS2e4E/F7vYQYGUPLO3IvUXQiTtS97tDj4Pl+8VywHTB7EQ+3URnH5DxUY0kmVDaOCk+MojsxXgKAJtRJ9KajSSI5puhi3aKH1BRsl+kPrMQQDn+nDMNf4acrKxwNmd8Ly9i13eltldtaAbDSWn06bJFFZm9086z+jXI5kweiBYNeOVxDL2MwiLsK2PSMMm/BBscQeuBCU7s1UTKiEpmgzKm0rd+yqmKokLPHfgptNW+Gkk57nkC04MOH0Y2Nc7LkivSUJMoqutIaqZkSYpCHvQjlisjDRsFrsRP53hSnT+wJeNZdCAkYI4+LbctzQfSP/j3gSXmv6pdBnSLq8s/O2cYhIkxupkK0mPxIqTylzcpRurH20OjcnBdXrBUhL8c9THO7E4asSyEMJEoNVs/3tkQbmOJdOOwICEZ1zlrygpUMU3cEDNLPd4sQKsN2RozYfS6XnPiqBgRWrWm0fkIlMO+/Uk1Q9rDMxdNUNoFQ328FqV5U3WNhN08UcxbukyKjoASLwCtPACsnU3YGsHV5Ft/GUBljguhfjDGPP9GSmGf9YQa7umNngJlV7smFx6PDJAtv/Gp9zi8gvufSezo3Kmk1v3v/hmR/YU4YRM2AaeaCY/u9Cg4Ncqlnr73q1AO7SWGgVdS2V8SX6uALjz/KkhP7tN6bST9I9EtuyIR0qqe9dqBH1h6sCBhkRoXbeQ+SiNszRdiBT1VsDpSiib519zHuFISUUb6hbL9je3yDr0EU+T0I7CZLzlDnBKxKoGOxpNZwvh27H5Eprwl0IjXo9kDUtmjMKgo8pS1SKXnme3Tb0Kn3GCrrPXuFbCrIAXZuT3iN5+QgwfFK7iCbhH/6y1mYaSiKfOPjfnds108t3AeRiNHnW2QXI8i+g8eyo8r2euO72P8mo5qJgCmZDc2RQEQXxCVA1VkvfBXJcJsaaLQNeqKVIVFiC0Rc0MJOhtqnElS1l9MkfGRrmtlhtxbYxYuhhZWExJiQR6/QTp5tp5p2UPZ085+XFmYR9nbJ4AyQP2cFag1NrdvQ3aPw7slJ/6kesdVnRM4uEBM9j4P14+nf8TLZzz20YFt5GNvmLF9eB9b7gOgTmZH54/etGjMRqqjgWtOiyqGsfXGzD3T5I8D2yvBmX2/hPmUl7QHDRmjI5wzLGHpTpIuF7odxO+5lntW5Xa+MXX8CgrhxGB+SWNA0SIJAidsR3RPqO/udHtEJ2JX2PNMFd3BMIdeyUe7SmmLM5ScKTCrsoRXg7X1MyY4+4eiIoXYp6ZghzhQYq0Uv/CmUFWMP9iJxQ763sLY03GPTaFGBdBQDXS9kr4peUV1XidifdmdEz0GBcKzmThQtYy/qbuYtZUhWPQVZaODiKK5NvBRt5FGSurt1Jr6l2hAU8BoXbucyUU+yIJeCW7EDWLhAb7luBBy5ohfKnXTA9YVt1QJ/UMc+p5SHwt+k6DqsKfF2Tf9PV0LWfGIZC3iITtS28g0Fm6AIWPPSKon1rzIml8xxF/0f+py8Ml4qKr//YkL/K9ODDIvu1tIs+bUFepCrKtKW4jFVHzavcALGYuM30ekXI7vz/P9IKIvImXpUusNAm+GITqTANIJLmYZuShaHLTTtyRFQxWC1Qzp6TwK2LpV0V51sSaPtQ69qE4RLf8Im91t0IsU1WDglZ2YQsd39UP2dgroqh7z3NJW369NS69qVgJ/nPNL45BfItKi5DNf+6GTT6ncXF5+vl7mk0EA+LgQcYKgoTN0UhvDUP6b49DA5UTLfJwD+EKzc5hKc17aZLbtRCwgJI65nrDCUxeEyDZokunaX4gyjRPqJGae6U6VgaR9ZPIng88sJvQ1aZ//Yft7AcNDXNmBLIlR36HFNZFFP/2OY3onIv/uJhtwz0axPUCuOrWbKeAasGzSpdQ+twwLHRIXf2ICc99jJYaTAxhF3/x206s6i2SuJpAJuDMG2hC54PyV055EiE/Wm8yAbzthIcBIQMVD+SGECrzfnut4NGb9KN8ndGbBW2wg6ZsBFwdxEZ5JUc8KllRMsZQGzT/STJuR2k9Io9+h8uq8ckBdRzBF/LfnYmjoywi5j32Lo21Xoxr3fEFMF1tbK3YCjLyL7ER4ReF+msnpDATB5CLF6tp94wgD50nBf8IResKhrtd6AxO7+Mn+7akjDgtzQsYdI2qEaOhX2si8oGFCzHUlCxPZTXzrxyL+EbmMAKGGoE4LN++/EvlfDzigx6XSNZ8OqqPmy0Lh0fdEyYsyLhijOHrOrTZYEj/Tw59mSst9pF9uYojb8XYZmjVy/qrE20gEbOvte9mBqSAjZ/XusdDPVZmAwZjgTn1BDkNyUU3szi1naK9rFI+acPkPtl4N5pHiqkSNl7/y28eUj3Rp7718EHJOxDCzR3NPSm6tUaCC+iePQdUiAHrWBOqFGxkYTm5swDigYUcop7dhTagib0utMxTpp6Znh0s9VvsLH7oynnpr9+plJaYhxYJ06moLIM3+Qb40jJ4xTMjmpRDFS3OPiIMiKp2H1T4ejhYXTJqsazA9Dfg6kbyTQ09QqW51V1h0jf1OtbkGwA/kFqyS5NLx/mTglDk6eoFxZP2iLqBr49hX9T8D5XRwWSZ5/ULlSCLEoiVyrTOgRRyMr7buHTWkPQDA64XBSpY8G9R4mW9mbASFUk0Nit1Ul53JtVqj17l0sGoo7CE++TDfDdYLvv7fwBMyA9Fvlaa6bd2pW6y/5e+zvO/jMpHXqZUO7KXxl/p2bVAoThXx0G1krrkme+NG7AEUSGhoNhhwRodpF6/p1wQdFFW1c1RPmzc4gss9g4/i4l8Dev+dMB5KoHwoT8mC6oGZlcH7lvEw0GNDBW3luoJLjWoUYCWeCYrurpAUv9JGQjkKAeThjZmELMjLSh4/Ersn+HNIg/lvaifBzoU3gFfP7l2YU9V4nDRal48im3vDtbCOuK1f4WjfQ2sPqGua5qVXYsd08Ylr+ufSHZB0wgnuLdaWK57BnWHHTmVxweoccVIdR7s8nIz0tiJr3ixJh2S4u66UfIh7VDVeOFgdcoTzB5jl60N601Kw4fPcC2xRDoONOAqcV6Gp4YC0CwbhZncJ0AF4GNMUT5qzktO6tzzR/iigqR22zUqNRmQmiDhg+abEFfmpIjKYihEsNbr4iRO43Je4v52Iaw/s895rr/DN3eZtLlVV44bqhQE/WTfwPzlFgxv2L4dhrP0NDyiXxEQ3tv5Dh6Xif7ocLJde0Ju4EWjdIoxTPRnIS+GuL8SwpdYq3/yc3nEP9JeBmlSPD7gpGnFdgq70xadN1vHsZV1r1vICh1znRUzyLYIKeR+bza93c2UVMEgwxGq1fgJjkuEb3f3AHVR6Y4VIEC0yh82CI7jn/Jt1jp9gKfN0739L8HG7Nat9VbLekdZI0NR2xM8mndQscvmNorRX9gv+SlcHQm/8tFD0gY9LZvjKX9LhbS+ZjSiGWPldD5Gai+xO5KNS/x+DclyKxX6gdBwapyT4hx/Lw/xamT7isqDPGZ0xlS9J6OwYR6qxGKJs4HTlFcMUCvXROp1I3ld0f9gPSCwFDk0+s2nLX2W8SxhgtN8ydYPqxQdKWAuQivSNkxlQ8AVOQ46vcC13VyLi/sS18gWwsHQUiSflbVADLH+zUQdKLlpcuNScPbyk0dZ1p7L9gN6J53Gh0ca6cfB5T2YiTJDUvlnTYxTGeggYfRn+3BYNMCGlGVqwhqHg4OEUOxL94gYQrmDMh+7Kl4z0JdmiyG1RyzhuSFFJq5BvKAup2RlHkWgP9bRKzTL+wBmbg0re1ow+Iovc87UKx8Qu0VM0Nj+B44+FOQVHmkZn0Z6uvm9/fkrRX+LvGvlOzPPFXuV++3RCjDW0BGd+OoveawTcBd0Ub9u8IoBI0IbRBjErPf1nmV42RpiRqdEWzdR30i1e5QdPpIOrYcZHJSUWzxWw2NHEy6Q0y/+Tw+ecTQH5X0tCtEmxuYSRIm6kMVUEfuyjCEREy+ZiW+PfnvehMj8w2NKjeWMEliRh4G+NBJWFy8OC8FVIrWEk2Iy4S2gw9XG8qHJSmc4NEPCLpesaPgaTXcogEx1Z63bkkwaiq87qVpbST65+N0P1sZWD0IE5Cd4/qX7WIhMCdqwC6VaMyxCGFCE3BAKwCRADNivA67EXO+LDxDXwmyZDYQJaOW4sxxxiZA+RagV8m8XyTJ7fRsiSwhrI2zqm2FDq7vXyFvTKZQFXtxDX+4NRw/e9SzelOLkwUNb9dZZTUnjvPpn4FixQ4ff9KasWG1se7+k5ge0y0zRtpKvsSqsHrue6/n73DGlwPtlbdSEJGrGa8TFGVz0hCldyhSLfQehzDCkyh+a7R5J6shvBseROJMxho0YvUfDYd2tfOBIF7PZ/MCBp7Qod7jmJ81jhEzFsRoGCfj64MKyX+SWZjYE0qADw+8mR7dJLLucqkshiHuCRmo9yQzCfBYwYe4GbJplq0uUNMObWqQ6aQktatXiL6I/B1q6PIAkF4bDfb5kBdVJaIzG5/Brb0zFl5P6o04213mQOZtFt1BQUoefSiq68Z5Q2+SfJ7zmHXvZJO6qBqiD0Lm6Kc6dYSvpeKELZR/dZSaSGEfhX5oRqyNWHxndD6T9XNetDgkhCdm+8brM0xSsY2x/PP9F7oMTQY6xPz+dnoZjtck3MRD5UzSASz63iNoNiLtU/tfrBbcSkr/PLBZ2B/i3krwfLfG83y6/n/kAnhvPrVGVlkjhcgTMLFsGcxyrV+4R50+UW8eqLqn01Baqtrzg0lkjtL8qKzn+nWXdNazv6XeEWDu66ymXbbc55T6FRXOHpKbQhIIY7kn0AXBhSVRYThkC5hf6cGzagERslKRumm/urwJVGZedClMoIXf0mxFH8Y7tbAZAlzrOhtP5WBZkzvmi+khmG/WRP8h+1A4tCVrWJtmtayFuteLd9Hl44zwysGK/+LnISQZ81VzoEd2NrauKQn1j2uygXpIq3gqiS55/iK/FObtbd1tkEqzM9jhTYk2Hn6ku8mHSqAfH5vdz3hHyeMOHLxsnYQLxpmbZamPEv8DBW/RV9UsvfbmvQh++6w+lYimf19WDBYPfkjLltYgij5EEW291SCbYsl7zw/9FJaopHoocEvDOaqdCMggwcLawayEX0QuhOjjGSoBD7eoShUJ6YDjkyA2KoqgCMTVFPqiJ+7WunnBEwZzEtBo7QB8OlFFwPy2Vt8hJofb+7VWQYmCkEjpfdjkH91grzkFDQ62SQggCiXsSaqvlBL/ccOPGpyR1fIkmP8JqKbqCKvUfXmnEEu7+irIot9qh1HcBeC5xqckVvK/bXbyfja+Lu1yX9mTUXu1tcCV07OU51MaOCa4lWFQ0orbj2R64TL+9FxI0daxbYMhwowIW4HXMFryO7bdE8XP48ImsiJOtvMfiMX1jdhYnJrhavQHkO1SduIgbu4s5deUCEdeIm708di/b0zOehppIYutuYDNtCn+P24BJr5FBxLu2mhzZESjz0YAjrhvlbl3uSgYQ1NSa1BTH1fYeASswnBUP9d3A/7lERU1B6r4TbVP1BaURY883fKoR/2r3X2f3qELAnSg4M/QD+9i4sbJjNd4Qzg6dVOCb4i+VkH5ouY9BalLp0RO0z2jlZ4e+zrmJNDrBzDF4l5+20wOdxWPKjMi84wkRbGC4V+LtjyrfPYNenxhomV6Q6GC67ijMbd94doBXJ1th4pB7ywrfpSF35tS1tE5QsRIGxwOt7S2RAjPFPjLEKFuBIeWQncHiJrGVrKeLrQktBU+WkHFZZv31/1N5xJFZ7uNaPGlM6U7WBNRZb780ZDf+yC1XdQGWwVS3kUPCv4m1bTW0MG01vGJ1AOOyQPkHWbuoofb+iiWkZ7+4yZzbs+JzhSSnA48VIGEruA4u6BKcDxdbDihINiy2a8Z83dF1bXGF68l6MjYKBprDnCMrElxQLBiObFanV4+6CZrVZ+psC7/4+SUTj91Ohfx78CwbFcNi4V53ft9AYr44cwLEGicrIuZ17zGVBRU2TqthMPtgidai68JNHY98Y5k/nd7/yaK7AGVx7eaYitlBTGTBj8VGAQdzqTqdKkGmZV1ijfsGCXaxIcTFycrRWjqlMZgxW/iMT8SMPqwWRbPH4OiYa0NG0aVqAuCUg6C7Droh1FxCqoj23ucI5r3mwjWU2m/2i2fr4EMklMP53BE/Akay/fG1m6gYUOeqwTp6Oy46hlXtEaL2HyuVkr4wZK5qClQs3X/S6DMVHdY5xleZE7zu0P6Lh3AzHLmUeBkwW3rP0WZ7oDOx0JwqNyTHwIf8L3cd044seE7I5ZsEPN1giWbrls/S1FW0ItfsNJL2SClWtBXE+/DjfQ4vzhm6EuMblU2TGNbEA3S0yhwoJbhxcMukvQ1bep5t7qP6fyOpmurE4NhUNgHgqz4U1tCsvFJNKMOZTVXD0dLN4YQN5mo+OYbImNVrKAoGhr+Z5RTvqEBSAjYnf/MG0j792wwk0etVEMdMYtBITtATNb4ytS1WPAu/bA5XGIuS8TrhLh8A4Fi0FYQb6JkMkRN7W9leYfjamhQexhtHmh3Cn5xIuBQ2SjQkHb6fIycdWCm50H7ETXuhbSb1W05IgureT2PTzu37nFC6K7+d+CpjoLw6VLvsa45/bOKNpplPIq03xmPlGKYUAkmwYmqelhXpEhXMyXKrL83ZSG38IVZRKIxjSrtCFczTAuGMd86/oBaiqPS+rznIxkV5iQ6bmNrpbE6DswMy4LCwgq+yTUHRgGZBlkmSTH5NYVstoLLzAz1dNQNv+AV+16ZxSV2j36ehJyJnNkQzRXphB+6FEWjYcMY4snF55JDFGyExAYQ6BRUIMMU6rPqpu1zlnpCrkWfc9ZOZ4utxH+FnMCoC2FBZNfiAA+yJ9nXKxJKGRrOE2MIzISku4pJY5h3Jka1XADykwCM8RNdruxPRiAbP6DGGjdYuvHF4l9wMiGJDYE9MW5WPrBqgIOJT7HU3HXYQTZy5KB+E8qCPp0oXDeZXdKI/4q/Zaqod/2lPZjawa0gmKbpFB38ENjtWIcoZNwbPhX0oTxHwk8PlWy7bYm0JhNlh3kJytyVKCdfWJ5yRq8O4xuFfeXQ6V/i0+33yrfgdWEwckQZfdYegd6XQh6JKS6Zjkx/WFtmk3dvAGwbDOhRlq9ScDzBnTE/fza/fLVfmz3ZATV1C45oMsczjzt53H7wokPAPM02Pzvdthb/OxFe8bj7XCLThSoBZ0U4zSMv6IB7BoLQjl49ZVXnNSlUJ71sDmomLkDYAdLQQcN24G0TmQ4dpzVnvNrk37FEojLqrCc7j8AgdLigHYOn+r3gdIxucwV8/3bDBWMM8+hOo5KiTdUPSsaRB+HcjzOJ0av0ndT8GlT7u7615p5o7LgG9HbNtcG+c2mSoTf2DAh1r1p6qs/Fv+bLy3K14uldFXPUecY5ZLDWqRXLgaWX4Pl7nbw/lh+15jp28pSVxoGkvbQaxZhV0MNF6q8EH1BiScWrFrvfOWIAQ8ane8uK+zBr3tAAwUeycRT0DXE8wKAYWVXN1VfrsgnTLTa6ddCzZSFtLjeeol1FxtAqvMhzVoIdBJwXW7pZA1V+gij3MDbqzWBbqArj+Xd+hbWaymkiwWXLdjdmp7YCWILLZwdws+4c4ENpVCrTP/iw2u8kiZOz2wuGNYX4GFE3GFh5yaQp4zs7Ktr4Dw1ctdxK45sLJYqucDv0tHjyXFxzLJ4T1kL5JbMSp2lEkKFqIAcdtrV7h0fWEHi7e5GXUK4BMrbZBQ4zKVE1JLmF3dZ9YQfRNr2cTohUD4bzggzuySwF/+JU0ZjT9Eq/H/Ts2Q+bV4waf1zrBNYlK4jmHAutRfVKNmUiqt+6vCYOZ/wy1wuMqNMGEqz91ztoJjWRDqUKACQTiM5hrKOWrSnMzQ6tChHasGQeQYHuBxVXUd0krfZLi0dNvOZA05Il46QBWtNscVO+p5DyZFozj4AIfhoP5RjG/0PEzm9aLe0I/tEa5PtAzvAP/AoDD0xmzZeggSf38j3ECB1JGWL/h8drpKvUkJDEfXFiWNelPyD/V2GL22nBPzWy3n16DssT1MM/ua9qhmDx72NPAtEzYhsyb4jAmjPZtOfzmBp3b8vhj1Df3Z3/np/HsavOo9ltKf2Dzf71WERueNRsiBUlMolTgk9gAAD+dlqUuWdOyAAG3epv9AwCb2oiLscRn+wIAAAAABFla")))
